Longitudinal study of schizophrenia symptoms with non-randomized treatment

This is a subset of a data set analyzed by Y. Korman on schizophrenia symptoms among patients who were assessed at 6 annual checkups. At each checkup, the type of drug currently prescribed by the treating physician was recorded.

Usage

data(Drugs)

Format

A data frame with 318 observations on 53 subjects each observed at 6 annual checkups with the following 11 variables.

age1

age in years at first checkup

yrsill1

years since initial diagnosis of schizophrenia at time of first checkup

pos

number of positive symptoms (symptoms that are not typically present in a normal subject

neg

number of negative symptoms (characteristics that are absent but normally present in normal subject

gen

other symptoms

total

sum of previous three

year

time of measurement (1 to 6)

drug

a factor with levels Atypical Clozapine Typical

status

a factor with levels Married Separated/Divorced Single

Sex

a factor with levels F M

Subj

a factor identifying each subject

Details

Source

References

Examples

data(Drugs)
## maybe str(Drugs) ; plot(Drugs) ...