Respiratory infections among Indonesian children. Longitudinal observational study with no treatment

275 preschool children in Indonesia were examined for up to 6 consecutive quarters for the present of respiratory infection. Also recorded presence of xerophthalmia, an eye condition indicating a vitamin A deficiency.

Usage

data(Indonesia)

Format

A data frame with 1194 observations on 275 children each measured on 1 to 6 occasions on the following 10 variables.

id

subject id

resp

indicator for a respiratory infection

age

age in months at time of measurement

baseline_age

age at start of program

xerophthalmia

indicator for the presence of xerophthalmia

sex

indicator for sex

height_for_age

height relative to age

stunted

indicator

time

time of measurement in quarters (3 month intervals)

season

1=Spring, 2=Summer, 3=Autumn, 4=Winter

Details

Desc: 275 preschool children in Indonesia were examined for up to 6 consecutive quarters for the present of respiratory infection. Also recorded presence of xerophthalmia, an eye condition indicating a vitamin A deficiency.

Source

Subset of cohort from Summer et al. (1984) reported in Diggle et al. (2002) 'Analysis of Longitudinal Data', 2nd ed. See: Zeger and Karim, JASA (1991)

References

Examples

data(Indonesia)
## maybe str(Indonesia) ; plot(Indonesia) ...