Stack data frames vertically

Stack data frames vertically on top of each other with columns identified by variable names. Matrices are coerced to data frames.

Usage

Rbind(..., verbose = FALSE, id = ".index")

Arguments

...

data frames or objects that can be coerced to data frames

Examples

zd1 <- data.frame(x = 1:3, y = 2, z = 3, .unique. = 'a')
zd2 <- data.frame(x = 1:4, y = 'y', w = 1, m = matrix(1:8, nrow = 4))
zm1 <- cbind(x = 1:3, y = 2, z = 3)
zm2 <- cbind(x = 1:4, y = 3, w = 1)
Rbind(zm1, zm2, zd1, zd1, zd2)
#>    x y .index  w  z .unique. m.1 m.2
#> 1  1 2      1 NA  3     <NA>  NA  NA
#> 2  2 2      1 NA  3     <NA>  NA  NA
#> 3  3 2      1 NA  3     <NA>  NA  NA
#> 4  1 3      2  1 NA     <NA>  NA  NA
#> 5  2 3      2  1 NA     <NA>  NA  NA
#> 6  3 3      2  1 NA     <NA>  NA  NA
#> 7  4 3      2  1 NA     <NA>  NA  NA
#> 8  1 2      3 NA  3        a  NA  NA
#> 9  2 2      3 NA  3        a  NA  NA
#> 10 3 2      3 NA  3        a  NA  NA
#> 11 1 2      4 NA  3        a  NA  NA
#> 12 2 2      4 NA  3        a  NA  NA
#> 13 3 2      4 NA  3        a  NA  NA
#> 14 1 y      5  1 NA     <NA>   1   5
#> 15 2 y      5  1 NA     <NA>   2   6
#> 16 3 y      5  1 NA     <NA>   3   7
#> 17 4 y      5  1 NA     <NA>   4   8
Rbind(Rbind(zm1, zm2), Rbind(zd1, zd2))
#>    x y .index  z  w .index. .unique. m.1 m.2
#> 1  1 2      1  3 NA       1     <NA>  NA  NA
#> 2  2 2      1  3 NA       1     <NA>  NA  NA
#> 3  3 2      1  3 NA       1     <NA>  NA  NA
#> 4  1 3      2 NA  1       1     <NA>  NA  NA
#> 5  2 3      2 NA  1       1     <NA>  NA  NA
#> 6  3 3      2 NA  1       1     <NA>  NA  NA
#> 7  4 3      2 NA  1       1     <NA>  NA  NA
#> 8  1 2      1  3 NA       2        a  NA  NA
#> 9  2 2      1  3 NA       2        a  NA  NA
#> 10 3 2      1  3 NA       2        a  NA  NA
#> 11 1 y      2 NA  1       2     <NA>   1   5
#> 12 2 y      2 NA  1       2     <NA>   2   6
#> 13 3 y      2 NA  1       2     <NA>   3   7
#> 14 4 y      2 NA  1       2     <NA>   4   8