Wrapper for read_excel in 'readxl' package

Facilitates reading excel file so each variable is read as a character without automatic coercion when it fails with 'read_excel'

Usage

Read_excel(
  path,
  sheet = 1,
  col_types = NULL,
  skip = 0,
  stringsAsFactors = default.stringsAsFactors(),
  ...
)

Arguments

path

path to xls/xlsx file

sheet

string or integer to identify sheet to read. Defaults to first sheet.

col_types

Either NULL to use guess from the spreadsheet made by 'read_excel' or a character vector containing 'blank', 'numeric', 'date' or 'text' which is recycled to match the number of columns in the spreadsheet. This behavior is different from that of 'read_excel' that does not recycle. If col_types is equal to 'text', all variables in the returned data frame have class character and the user needs to handle conversions manually.

skip

Number of rows to skip before reading any data.

stringsAsFactors

defaults to 'default.stringsAsFactors()'

col_names

Either TRUE to use first row as as column names, FALSE to number sequentially from X1 to Xn, or a character vector giving a name for each column.

na

Missing value. By default readxl converts blank cells to missing data. Set this value if you have used a sentinel value for missing values.

Value

a data frame. The value returned by readxl::read_excel

Details

read_excel, in the 'readxl' package, guesses the type of each column based on some initial number of values. If non-numeric values first occur quite far into the spread sheet, read_excel will misguess the type and the non-numeric values will be returned as missing when coerced to numeric. Using the argument "col_types = 'text'" in read_excel does not work, since read_excel expects the length of 'col_types' to equal the number of columns. 'Read_excel' first counts the number number of columns and recycles the 'col_types' argument so it has the right length.

See also

Examples