Transform output of a Wald test into a data frame

The generic method is included in the possibly false hope that it will ensure that the method in this package is used.

The generic method is included in the possibly false hope that it will ensure that the method in this package is used.

Usage

# S3 method for class 'wald'
as.data.frame(obj, se = 2, digits = 3, sep = "", which = 1, ...)

# S3 method for class 'wald'
as.data.frame(obj, se = 2, digits = 3, sep = "", which = 1, ...)

Arguments

obj

a 'wald' object

se

a vector with the multiples of standard error used to generate lower and upper limits. 'names(se)' is appended to 'L' and 'U' to label the variables.

which

selects elements of 'obj' to turn to a data.frame.

Value

A data frame with estimated coefficient, standard error, and, optionally, upper and lower limits and the variables included the 'data' element of 'obj' if present. If length(which) > 1, the returned object is a list of data frames.

A data frame with estimated coefficient, standard error, and, optionally, upper and lower limits and the variables included the 'data' element of 'obj' if present. The 'wald' object is returned as the 'wald' attribute so p-values, for example, can be obtained with attr(ret, 'wald')$estimate$p-value If length(which) > 1, the returned object is a list of data frames.

Examples

# coming soon!
# coming soon!