Conjugate complement of span(X) in span(Z) with respect to inner product ip

Conjugate complement of span(X) in span(Z) with respect to inner product ip

Usage

cc_svd(X, Z = NULL, ip = NULL, tol = 1e-14)

cc_qr(X, Z = diag(NROW(X)), ip = diag(NROW(X)), tol = 1e-14)

Arguments

X

matrix defining space whose complement is computed. Not necessarily of full column rank

Z

matrix defining space within which the complement is computed. Should be of full column rank. Default: diag(nrow(X))

ip

positive definite matrix defining inner product with respect to which complement is computed. Default: diag(nrow(X))

tol

tolerance (default 1e-14)

Functions

cc_qr(): Conjugate complement using QR algorithm

Examples

#' cc_svd(cbind(1:3))
cc_svd(cbind(1:3), cbind(1:3,1))
#>      [,1]
#> [1,]    0
#> [2,]    0
#> [3,]    0
cc_svd(cbind(1:3), cbind(3*c(1,-2,1), 1:3,1))
#>      [,1]
#> [1,]    0
#> [2,]    0
#> [3,]    0
# challenge with large numbers
K <- cc_svd(cbind(1,1:10,(1:10)^2))
K
#>       [,1]
#>  [1,]    0
#>  [2,]    0
#>  [3,]    0
#>  [4,]    0
#>  [5,]    0
#>  [6,]    0
#>  [7,]    0
#>  [8,]    0
#>  [9,]    0
#> [10,]    0
ret <- list()
for(i in 1:20){
  cbind(1,(1:10-10^i)^2,1:10) %>% 
    { list(svd = ccor(cc_svd(.), K) -1, qr =ccor(cc_qr(.), K) -1)
    }  %>% lapply(abs) %>% 
    {list(med=lapply(.,median), ave = lapply(.,mean))} %>% unlist->ret[[i]]
}
#> Error in cancor(x, y, F, F): 'x' has rank 0
ret <- do.call(rbind, ret)
ret
#> NULL
library(lattice)
library(latticeExtra)
xyplot(ts(log2(ret)),  scales = list(y = 'same'), ylab = 'cancorr: log base 2 of deviation from 1')
#> Error in log2(ret): non-numeric argument to mathematical function
#
# challenge with small numbers
#
i <- 1
K <- cc_svd(cbind(1,1:10,(1:10 == 1)* 10^{-i}))
K
#>       [,1]
#>  [1,]    0
#>  [2,]    0
#>  [3,]    0
#>  [4,]    0
#>  [5,]    0
#>  [6,]    0
#>  [7,]    0
#>  [8,]    0
#>  [9,]    0
#> [10,]    0
ret <- list()
for(i in 1:30){
  cbind(1,1:10,(1:10 == 1)* 10^{-i}) %>% 
    { list(svd = ccor(cc_svd(.), K) -1, qr =ccor(cc_qr(.), K) -1)
    }  %>% lapply(abs) %>% 
    {list(med=lapply(.,median), ave = lapply(.,mean))} %>% unlist->ret[[i]]
}
#> Error in cancor(x, y, F, F): 'x' has rank 0
ret <- do.call(rbind, ret)
ret
#> NULL

xyplot(ts(log2(ret)),  scales = list(y = 'same'), ylab = 'cancorr: log base 2 of deviation from 1')
#> Error in log2(ret): non-numeric argument to mathematical function

#
# challenge with small numbers added to 1
#
i <- 1
K <- cc_svd(cbind(1,1:10,1+(1:10 == 1)* 10^{-i}, (1:10 != 1)))
K
#>                [,1]          [,2]
#>  [1,]  1.135974e-17 -7.306532e-17
#>  [2,] -4.175853e-01  3.039922e-01
#>  [3,]  2.771731e-01 -2.844012e-01
#>  [4,]  7.725156e-02 -7.947974e-02
#>  [5,]  6.436167e-01 -1.015384e-01
#>  [6,] -3.935856e-01 -1.235972e-01
#>  [7,] -3.617429e-01 -1.456559e-01
#>  [8,]  2.046223e-01  8.322854e-01
#>  [9,] -3.079631e-02 -1.897733e-01
#> [10,]  1.046367e-03 -2.118320e-01
#> attr(,"d")
#> [1] 2.021425e+01 2.237956e+00 7.653573e-01 6.970080e-18
ret <- list()
for(i in 1:30){
  cbind(1,1:10, 1 + (1:10 == 1)* 10^{-i}) %>% 
    { list(svd = ccor(cc_svd(.), K) -1, qr =ccor(cc_qr(.), K) -1)
    }  %>% lapply(abs) %>% 
    {list(med=lapply(.,median), ave = lapply(.,mean))} %>% unlist->ret[[i]]
}
#> Error in cancor(x, y, F, F): 'x' has rank 0
ret <- do.call(rbind, ret)
ret
#> NULL