Create a contextual variable for regression

cvar and See also dvar are designed to be used in regression formulas to create a contextual mean of a cluster-varying variable and a 'centered-within-groups' version.

cvar and See also dvar are designed to be used in regression formulas to create a contextual mean of a cluster-varying variable and a 'centered-within-groups' version.

cvar and dvar are designed to be used in regression formulas to create a contextual mean of a cluster-varying variable and a 'centered-within-groups' version. See also cvar.

Usage

cvar(x, id, all, na.rm, ...)

# S3 method for class 'factor'
cvar(x, id, all = FALSE, na.rm = TRUE, ...)

# Default S3 method
cvar(x, id, all, na.rm = TRUE, ...)

cvar(x, id, all, na.rm, ...)

# S3 method for class 'factor'
cvar(x, id, all = FALSE, na.rm = TRUE, ...)

# Default S3 method
cvar(x, id, all, na.rm = TRUE, ...)

dvar(x, id, all, na.rm, ...)

Arguments

x

variable to be centered or residualized within groups. If x is a factor, cvar and dvar return matrices whose columns are named consistently with the names of coding variables for factors.

id

identifies clusters

all

(default FALSE) if TRUE cvar.factor returns the columns means of an incidence matrix including the first level. Otherwise, the first level is dropped for use in a linear model.

na.rm

(default TRUE) whether to drop missing values

FUN

(default mean) function to be applied to create contextual values

Methods (by class)

cvar(factor): method for class 'factor'

cvar(default): default method

cvar(factor): method for class 'factor'

cvar(default): default method

Examples

if (FALSE) { # \dontrun{
dd <- data.frame(x= 1:100, id = rep( LETTERS[1:10], each = 10))
dd$a <- factor(sample( c('a','b','c'), 100, replace = T))
dd$y <- dd$x + rep(rnorm(10), each = 10) + rnorm(100) + as.numeric(dd$a)
library(nlme)
fit <- lme( y ~ x + cvar(x,id), dd, random = ~ 1 + dvar(x,id) | id)
anova( fit , type = 'm')
# The output of 'anova' can be used to test whether a contextual variable
# should be included in the model
} # }
if (FALSE) { # \dontrun{
dd <- data.frame(x= 1:100, id = rep( LETTERS[1:10], each = 10))
dd$a <- factor(sample( c('a','b','c'), 100, replace = T))
dd$y <- dd$x + rep(rnorm(10), each = 10) + rnorm(100) + as.numeric(dd$a)
library(nlme)
fit <- lme( y ~ x + cvar(x,id), dd, random = ~ 1 + dvar(x,id) | id)
anova( fit , type = 'm')
# The output of 'anova' can be used to test whether a contextual variable
# should be included in the model
} # }
if (FALSE) { # \dontrun{
dd <- data.frame(x= 1:100, id = rep( LETTERS[1:10], each = 10))
dd$a <- factor(sample( c('a','b','c'), 100, replace = T))
dd$y <- dd$x + rep(rnorm(10), each = 10) + rnorm(100) + as.numeric(dd$a)
library(nlme)
fit <- lme( y ~ x + cvar(x,id), dd, random = ~ 1 + dvar(x,id) | id)
anova( fit , type = 'm')
# The output of 'anova' can be used to test whether a contextual variable
# should be included in the model
} # }