Drop each row or cluster from a fitted object

Drop each row or each cluster in a fitted object and collect the resulting linear estimates. Return the drop one estimates along with variable from original data that are invariant within clusters, as well as the maximum value of DFBETAS for each cluster dropped.

Usage

dropone(
  fit,
  form = NULL,
  FUN = if (inherits(fit, "lme")) fixef else coef,
  data = getData(fit),
  ...
)

parDropone(
  fit,
  form = NULL,
  FUN = if (inherits(fit, "lme")) fixef else coef,
  data = getData(fit),
  mc.cores = detectCores(),
  ...
)

Arguments

fit

a fitted object with an update method and a getData method.

form

a formula evaluated in the data frame for 'fit' that defines clusters to be dropped one at a time. If NULL, the default, each row is dropped successively.

FUN

function used to extract coefficients. Default: fixef for class 'lme' and coef otherwise.

data

data frame in which to evaluate the refitted object. Default: 'getData(fit)'.

mc.cores

number of cores to use. Default: parallel::detectCores() \(\exp^n\)

Functions

parDropone(): parallel computing version of dropone using the parallel package