Get information on fixed effects from a model object

getFix extracts coefficients, covariance matrix and degrees of freedom from model objects. Its main purpose is to extract information need by the wald function. To extend the wald function to a new class of objects, it is merely necessary to write a method for getFix.

getFix extracts coefficients, covariance matrix and degrees of freedom from model objects. Its main purpose is to extract information need by the wald function. To extend the wald function to a new class of objects, it is merely necessary to write a method for getFix.

Usage

getFix(fit, ...)

# S3 method for class 'multinom'
getFix(fit, ...)

# S3 method for class 'lm'
getFix(fit, robust = FALSE, type = "HC3", ...)

# S3 method for class 'glm'
getFix(fit, robust = FALSE, ...)

# S3 method for class 'lme'
getFix(fit, robust = FALSE, ...)

# S3 method for class 'gls'
getFix(fit, robust = FALSE, ...)

# S3 method for class 'lmer'
getFix(fit, robust = FALSE, ...)

# S3 method for class 'glmer'
getFix(fit, robust = FALSE, ...)

# S3 method for class 'lmerMod'
getFix(fit, robust = FALSE, ...)

# S3 method for class 'glmerMod'
getFix(fit, robust = FALSE, ...)

# S3 method for class 'mer'
getFix(fit, robust = FALSE, ...)

# S3 method for class 'zeroinfl'
getFix(fit, robust = FALSE, ...)

# S3 method for class 'mipo'
getFix(fit, robust = FALSE, ...)

# S3 method for class 'MCMCglmm'
getFix(fit, robust = FALSE, ...)

# S3 method for class 'stanfit'
getFix(fit, pars, include = TRUE, robust = FALSE, ...)

# Default S3 method
getFix(fit, ...)

getFix(fit, ...)

# S3 method for class 'multinom'
getFix(fit, ...)

# S3 method for class 'lm'
getFix(fit, robust = FALSE, type = "HC3", ...)

# S3 method for class 'glm'
getFix(fit, robust = FALSE, ...)

# S3 method for class 'lme'
getFix(fit, robust = FALSE, ...)

# S3 method for class 'gls'
getFix(fit, robust = FALSE, ...)

# S3 method for class 'lmer'
getFix(fit, robust = FALSE, ...)

# S3 method for class 'glmerMod'
getFix(fit, robust = FALSE, ...)

# S3 method for class 'glmer'
getFix(fit, robust = FALSE, ...)

# S3 method for class 'mer'
getFix(fit, robust = FALSE, ...)

# S3 method for class 'zeroinfl'
getFix(fit, robust = FALSE, ...)

# S3 method for class 'mipo'
getFix(fit, robust = FALSE, ...)

# S3 method for class 'MCMCglmm'
getFix(fit, robust = FALSE, ...)

# S3 method for class 'stanfit'
getFix(fit, pars, include = TRUE, robust = FALSE, ...)

# S3 method for class 'lmerMod'
getFix(fit, robust = FALSE, ...)

# S3 method for class 'glmerMod'
getFix(fit, robust = FALSE, ...)

# Default S3 method
getFix(fit, ...)

Arguments

fit

A fitted model object

...

Other arguments unused

Value

Returns a list with the following components:

fixedFixed effect parameter estimates

vcovCovariance matrix of the parameters

dfdenominator degrees of freedom for each effect

Returns a list with the following components:

fixedFixed effect parameter estimates

vcovCovariance matrix of the parameters

dfdenominator degrees of freedom for each effect

Details

Extending the wald function to a new class of objects only requires a getFix method for the new class.

Extending the wald function to a new class of objects only requires a getFix method for the new class.

Methods (by class)

getFix(multinom): method for multinom objects in package nnet

getFix(lm): method for lm objects

getFix(glm): method for glm objects

getFix(lme): method for lme objects in the nlme package

getFix(gls): method for gls objects in the nlme package

getFix(lmer): method for lmer objects in the lme4 package

getFix(glmer): method for glmer objects in the lme4 package

getFix(lmerMod): method for lmerMod objects in the lme4 package

getFix(glmerMod): method for glmerMod objects in the lme4 package

getFix(mer): method for mer objects in the lme4 package

getFix(zeroinfl): method for zeroinfl objects in the pscl?? package

getFix(mipo): method for mipo objects in the mice package

getFix(MCMCglmm): method for MCMCglmm objects in the MCMCglmm package

getFix(stanfit): method for stanfit objects in the rstan package

getFix(default): print message if getFix id used for a class for which a method has not been written

getFix(multinom): method for multinom objects in package nnet

getFix(lm): method for lm objects

getFix(glm): method for glm objects

getFix(lme): method for lme objects in the nlme package

getFix(gls): method for gls objects in the nlme package

getFix(lmer): method for lmer objects in the lme4 package

getFix(glmerMod): method for glmerMod objects in the lme4 package

getFix(glmer): method for glmer objects in the lme4 package

getFix(mer): method for mer objects in the lme4 package

getFix(zeroinfl): method for zeroinfl objects in the pscl?? package

getFix(mipo): method for mipo objects in the mice package

getFix(MCMCglmm): method for MCMCglmm objects in the MCMCglmm package

getFix(stanfit): method for stanfit objects in the rstan package

getFix(lmerMod): method for lmerMod objects in the lme4 package

getFix(glmerMod): method for glmerMod objects in the lme4 package

getFix(default): print message if getFix id used for a class for which a method has not been written

See also

Author

Georges Monette

Examples

library(nlme)
fit <- lme( mathach ~ (ses + I(ses^2)) * Sex, hs, random = ~ 1 + ses| school)
getFix(fit)
#> $fixed
#>      (Intercept)              ses         I(ses^2)          SexMale 
#>      12.07087351       2.61144284       0.26459124       1.09633850 
#>      ses:SexMale I(ses^2):SexMale 
#>      -0.33162835       0.07429928 
#> 
#> $vcov
#>                   (Intercept)          ses     I(ses^2)      SexMale
#> (Intercept)       0.164938076 -0.013979664 -0.043769531 -0.073094075
#> ses              -0.013979664  0.087909509  0.006791532  0.003896157
#> I(ses^2)         -0.043769531  0.006791532  0.080379764  0.042284847
#> SexMale          -0.073094075  0.003896157  0.042284847  0.161068134
#> ses:SexMale       0.006482282 -0.073446962 -0.009787746 -0.017940034
#> I(ses^2):SexMale  0.043544791 -0.008234621 -0.075966113 -0.094188126
#>                   ses:SexMale I(ses^2):SexMale
#> (Intercept)       0.006482282      0.043544791
#> ses              -0.073446962     -0.008234621
#> I(ses^2)         -0.009787746     -0.075966113
#> SexMale          -0.017940034     -0.094188126
#> ses:SexMale       0.158575843      0.025723617
#> I(ses^2):SexMale  0.025723617      0.163399621
#> 
#> $df
#>      (Intercept)              ses         I(ses^2)          SexMale 
#>             1932             1932             1932             1932 
#>      ses:SexMale I(ses^2):SexMale 
#>             1932             1932 
#> 

data(Prestige, package="car")
#> Warning: data set ‘Prestige’ not found
mod.prestige <- lm(prestige ~ education + income, data=Prestige)
#> Error in eval(mf, parent.frame()): object 'Prestige' not found
getFix(mod.prestige)
#> Error in eval(expr, envir, enclos): object 'mod.prestige' not found

library(nlme)
fit <- lme( mathach ~ (ses + I(ses^2)) * Sex, hs, random = ~ 1 + ses| school)
getFix(fit)
#> $fixed
#>      (Intercept)              ses         I(ses^2)          SexMale 
#>      12.07087351       2.61144284       0.26459124       1.09633850 
#>      ses:SexMale I(ses^2):SexMale 
#>      -0.33162835       0.07429928 
#> 
#> $vcov
#>                   (Intercept)          ses     I(ses^2)      SexMale
#> (Intercept)       0.164938076 -0.013979664 -0.043769531 -0.073094075
#> ses              -0.013979664  0.087909509  0.006791532  0.003896157
#> I(ses^2)         -0.043769531  0.006791532  0.080379764  0.042284847
#> SexMale          -0.073094075  0.003896157  0.042284847  0.161068134
#> ses:SexMale       0.006482282 -0.073446962 -0.009787746 -0.017940034
#> I(ses^2):SexMale  0.043544791 -0.008234621 -0.075966113 -0.094188126
#>                   ses:SexMale I(ses^2):SexMale
#> (Intercept)       0.006482282      0.043544791
#> ses              -0.073446962     -0.008234621
#> I(ses^2)         -0.009787746     -0.075966113
#> SexMale          -0.017940034     -0.094188126
#> ses:SexMale       0.158575843      0.025723617
#> I(ses^2):SexMale  0.025723617      0.163399621
#> 
#> $df
#>      (Intercept)              ses         I(ses^2)          SexMale 
#>             1932             1932             1932             1932 
#>      ses:SexMale I(ses^2):SexMale 
#>             1932             1932 
#> 

data(Prestige, package="car")
#> Warning: data set ‘Prestige’ not found
mod.prestige <- lm(prestige ~ education + income, data=Prestige)
#> Error in eval(mf, parent.frame()): object 'Prestige' not found
getFix(mod.prestige)
#> Error in eval(expr, envir, enclos): object 'mod.prestige' not found