Extract or remove a substring

Extract of remove a substring matching a regular expression.

Usage

getex(x, pattern, ...)

Arguments

x

object from which strings are extracted. Will be coerced to a character by regexpr.

pattern

a regular expression whose matches in x are returned.

...

other arguments passed to regexpr.

Value

getex returns the substrings matched by pattern in x. getex_ returns the portion of x that is not matched.

Details

getex and getex_ can work as a pair to gradually extract information from a string.

Examples

library(magrittr)
# The following could be done with 'strsplit' but 'getex' can handle messier cases
progs <- c('SC BSc Honours Biology', 'Arts BA Ordinary Economics')
faculty <- progs  %>%  getex('^[^ ]+') %>% toupper
#> Error in getex(., "^[^ ]+"): could not find function "getex"
progs <- progs  %>% getex_('^[^ ]+ ')
#> Error in getex_(., "^[^ ]+ "): could not find function "getex_"
degree <- progs %>% getex('^[^ ]+')
#> Error in getex(., "^[^ ]+"): could not find function "getex"