General parametric regression splines with variable degrees and smoothness

This function creates functions (closures) that implement general polynomial splines with possibly different degrees in each interval and different orders of smoothness at each knot, including the possibility of allowing a discontinuity at a knot.

For transitional reasons with the forthcoming gspline package, Gspline is a synonym for gspline.

Usage

gspline(
  knots,
  degree = 3,
  smoothness = pmax(pmin(degree[-1], degree[-length(degree)]) - 1, 0),
  intercept = 0,
  constraints = NULL,
  estimates = NULL,
  periodic = FALSE,
  tolerance = 1e-14,
  debug = FALSE
)

# S3 method for class 'gspline'
print(
  x,
  show = c("knots", "degree", "smoothness", "G", "constraint_mat", "estimate_mat"),
  ...
)

# S3 method for class 'gspline_matrix'
print(x, ...)

knots_.gspline(sp)

Gspline(
  knots,
  degree = 3,
  smoothness = pmax(pmin(degree[-1], degree[-length(degree)]) - 1, 0),
  intercept = 0,
  constraints = NULL,
  estimates = NULL,
  periodic = FALSE,
  tolerance = 1e-14,
  debug = FALSE
)

Arguments

knots

vector of knots

degree

vector giving the degree of the spline in each interval. Note the number of intervals is equal to the number of knots + 1. A value of 0 corresponds to a constant in the interval.

smoothness

vector with the degree of smoothness at each knot (0 = continuity, 1 = smooth with continuous first derivative, 2 = continuous second derivative, etc. The value -1 allows a discontinuity at the knot. Alternative, a list each of whose elements specifies the derivatives that must be continuous at each knot. This allow the set of continuous derivatives to skip a value. For example, to specify that a function can have a discontinuity but must have the same limiting slope and second derivative (curvature) on either side of the discontinuity, the smoothness vector for the corresponding knot would have the value c(1,2), thus leaving the value (denoted by 0) unconstrained.

constraints

provides a vector or matrix specifying additional linear contraints on the 'full' parametrization consisting of blocks of polynomials of degree equal to max(degree) in each of the length(knots)+1 intervals of the spline. See below for examples of a spline that is 0 outside of its boundary knots.

estimates

provides a vector or matrix specifying additional linear combination of the parameters in the 'full' parametrization that should be estimated by estimated coefficients of the spline.

periodic

if TRUE a periodic spline is generated on the base interval 0,max(knots). A constraint is generated so that the coefficients generate, in effect, the same fitted values and derivatives to the right of max(knots) as they do to the right of 0. Note that all knots must be strictly positive. Since the knot at 0 corresponds to the knot at 'max(knots)', the knot at 0 must not be explicitly specified. Note also that the degree in the interval to the right of 'max(knots)' must be equal to that in the interval to the left of 'min(knots)'. If these are specified to be equal, gspline uses the maximum for both intervalsl. The continuity contraints associated with the last knot are, in effect, continuity constraints at 0.

x

(parameter of the closure created by gspline) value(s) where the spline is evaluated. This is the only parameter used when generating portions of a model matrix.

signif

number of significant digits used to label coefficients

exclude

(not yet implemented) number of leading columns to drop from spline matrix: 0: excludes the intercept column, 1: excludes the linear term as well. Terms that are excluded from the spline matrix can be modeled explicitly.

a

(not yet implemented) a 'periodic' knot at which a spline repeats itself

n

(not yet implemented) the maximal 'order', i.e. maximal degree + 1, of a periodic spline

D

(parameter of the closure created by gspline) value(s) the order of derivatives to be estimated when creating a linear hypothesis matrix. The default is D=0 corresponding to the value of the spline.

limit

(parameter of the closure created by gspline) specifies whether a derivative or value should be estimated as as limit from the left limit = -1, from the right, limit = -1, or the difference of the two limits, limit = -0. The default is limit = -1.

Value

gspline returns a closure that creates portions of model matrices or of linear hypothesis matrices for Wald test for a general polynomial spline.

Details

The functions created by gspline can also be used to generate portions of a linear hypothesis matrix that can be used with a Wald test to estimate features of the spline such as derivatives of various orders, discontinuities, etc.

Many polynomial regression splines can be generated by 'plus' functions although the resulting basis for the spline may be ill conditioned. For example, a 'quadratic spline' (a spline that is quadratic in each interval with and smooth with a first derivative at each knot) with knots at 1 and 3 can be fitted with:

plus <- function(x,y) ifelse(x>0,y,0)
lm(y ~ x + I(x^2) + plus(x-1,(x-1)^2) + plus(x-3,(x-3)^2))

All 'standard' polynomial splines with the same degree in each interval and continuity of order one less than the degree at each knot can be constructed in this fashion. A convenient aspect of this parametrization of the spline is that the estimated coefficients have a simple interpretation. The coefficients for 'plus' terms represent the 'saltus' (jump) in the value of a coefficient at the knot. Testing whether the true value of the saltus in a coefficient is 0 is equivalent to a test for the need to retain the corresponding knot.

This approach does not work for some more complex splines with different degrees or different orders of continuity at the knots.

Many techniques for fitting splines generate a basis for the spline (columns of the design matrix) that has good numerical properties but whose coefficients do not have a simple interpretation.

The gspline function generates functions representing splines with arbitrary degrees in each interval and arbitrary smoothness at each knot. Any set of orders of derivatives can be contrained to be continuous at any particular knot. The parametrization produces coefficients that have a simple interpretation. For a spline of degree p at x = 0, coefficients correspond to the 1st, 2nd, ... pth derivative at 0. Additional coefficients correspond to each free discontinuities at each knot.

A disadvantage of spline functions generated by gspline is that the spline basis may be poorly conditioned. The impact of this problem can be mitigated by rescaling the x variable so that it has an appropriate range. For example, with a spline whose highest degree is cubic, ensuring that x has a range between -10 and 10 should avert numerical problems.

Let sp denote a spline matrix created by gspline. For example sp <- gspline(knots = c(-1,0,2), degree = c(2,3,3,2), smoothness = c(1,-1,1)) creates a spline function sp that can be used in linear formulas to fit a polynomial spline with knots at -1, 0 and 2, with degrees 2, 3, 3, and 2 in the four intervals bounded by the knots and with "C 1" continuity (continuous first derivative) at the knots -1 and 2, and a possible discontinuity in value at the knot 0.

The linear formula to fit this spline has the form y ~ 1 + sp(x) which can be augmented with other regressors in the usual way.

Called with a single argument (e.g. sp(x)) sp returns a matrix of regressors and can be used in linear model formulas, e.g y ~ sp(x), y ~ A * sp(x) or y ~ A / sp(x) -1 if A is a factor to fit separate splines within each level of A.

sp can be called with additional argument to geneate portions of linear hypothesis matrices that can be used test or estimate various values or derivatives of the spline. For example, sp(c(-1,-1,-1,0,2), D = c(2, 2, 2, 0, 1), limit = c(-1, 1, 0, 0, 1)) will return a matrix of coefficients to estimate, respectively, the limits from the left and from the right of the second derivative at -1, the difference (jump or saltus) in this derivative at -1, the size of the discontinuity in the value of the function at 0, and the first derivative at 1, which being continuous there has the same value whether a limit is taken from the right or from the left.

A function to fit a cubic spline with knots at 5 and 10 is generated with sp <- gspline(knots = c(5, 10), degree = c(3, 3, 3), smoothness = c(2, 2)), or more briefly: sp <- gspline(c(5, 10), 3, 2) since the degree and smoothness arguments are recycled as required by the number of knots.

A natural cubic spline with knots at 5 and 10 and boundary knots at 0 and 20 can be created with sp <- gspline(c(0, 5, 10, 20), c(1, 3, 3, 3, 1), c(2, 2, 2, 2)).

A step function with steps at 0, 1 and 2 would have the form: sp <- gspline(c(0, 1, 2), 0, -1).

Warning

The estimands generated by spline functions created by gspline are designed so the coefficients are interpretable as changes in derivatives at knots. The resulting matrix is not designed to have optimal numerical properties.

Examples