Articial data with classical outliers.

Articial data on health as predicted by height and weight with outliers to show the effect of different types of outliers.

Usage

data(hw)

data(hwoutliers)

Format

A data frame with 4 datasets, indexed by Type, each with 15 'good' observations. The last 3 datasets eadh contain, in addition to the 'good' data, an outlier.

Weight

a measure of body weight

Height

a measure of height

Health

an index of health: higher is better

Type

numerical type, 0, 1, 2 or 3

Outlier

a factor with levels none Type 1 Type 2 Type 3

Details

hw has four datasets identified by Outlier indicating the 'type' of outlier depicted in each dataset: none, Type 1: large residual for y with typical value of x, Type 2: small residual for y with atypical value of x, and Type 3: atypical x and poorly fitting y. hwoutliers contains the rows of 'acceptable' data and one row for each of the three types of outlier. This format lends itself better to 3-D plots. Note that this nomenclature for outlier is not widely adopted!

Source

References

Examples

data(hw)
## maybe str(hw) ; plot(hw) ...