Longitudinal study of IQ after traumatic brain injuries

A subset of data gathered on verbal and performance IQ of patients recovering from coma after traumatic brain injuries iq is a subset of the original data, iqsim is simulated set of 10 hypothetical observations on one subject.

Usage

iq

Format

A longitudinal with 331 observations 200 subjects measured on a varying number of occasions (ranging from 1 to 5) on the following 7 variables.

DAYSPC

time of measurement in days post recovery from coma

DCOMA

duration of coma rounded to nearest day

SEX

a factor with levels Female Male

AgeIn

age at time of injury

ID

identifying subjects

PIQ

performance (or mathematical) IQ

VIQ

verbal IQ

References

Wong, Monette, Wiener COMPLETE