Post-stratification jacknife estimates of standard error

Jacknife estimates of the weighted mean with a stratified sample. Weights are adjusted to take the change in n into account when obtaining the jacknife estimate dropping an observation within a stratum.

Usage

jk_wtd_means(x, by, w, check = TRUE, FUN = wtd_mean)

Arguments

x

numerical vector

by

stratification variable or list of stratification variable

w

vector of weights, e.g. Horvitz-Thomson weights or renormed version

check

(default: TRUE) check that 'w' is constant within levels of 'by'

FUN

(default: wtd_mean) function used to obtain estimate

See also