Print array using kableExtra

Print a higher-dimensional array with subheadings for the levels defined by the third and higher dimensions of the array.

Usage

kable_array(a, caption = "", ...)

Arguments

a

array

caption

an optional caption for the printed table

...

functions from kable extra and their arguments

Value

hmtl code to include in a 'asis' chunk

Examples

kable_array(Titanic)
#> Loading required package: kableExtra
#> [1] 1
#> <table class="table table" style="width: auto !important; margin-left: auto; margin-right: auto; margin-left: auto; margin-right: auto;">
#>  <thead>
#>   <tr>
#>    <th style="text-align:left;position: sticky; top:0; background-color: #FFFFFF;">  </th>
#>    <th style="text-align:right;position: sticky; top:0; background-color: #FFFFFF;"> Male </th>
#>    <th style="text-align:right;position: sticky; top:0; background-color: #FFFFFF;"> Female </th>
#>   </tr>
#>  </thead>
#> <tbody>
#>   <tr grouplength="4"><td colspan="3" style="border-bottom: 1px solid;"><strong>Survived: No / Age: Child</strong></td></tr>
#> <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 1st </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 2nd </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 3rd </td>
#>    <td style="text-align:right;"> 35 </td>
#>    <td style="text-align:right;"> 17 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> Crew </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>   </tr>
#>   <tr grouplength="4"><td colspan="3" style="border-bottom: 1px solid;"><strong>Survived: No / Age: Adult</strong></td></tr>
#> <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 1st </td>
#>    <td style="text-align:right;"> 118 </td>
#>    <td style="text-align:right;"> 4 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 2nd </td>
#>    <td style="text-align:right;"> 154 </td>
#>    <td style="text-align:right;"> 13 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 3rd </td>
#>    <td style="text-align:right;"> 387 </td>
#>    <td style="text-align:right;"> 89 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> Crew </td>
#>    <td style="text-align:right;"> 670 </td>
#>    <td style="text-align:right;"> 3 </td>
#>   </tr>
#>   <tr grouplength="4"><td colspan="3" style="border-bottom: 1px solid;"><strong>Survived: Yes / Age: Child</strong></td></tr>
#> <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 1st </td>
#>    <td style="text-align:right;"> 5 </td>
#>    <td style="text-align:right;"> 1 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 2nd </td>
#>    <td style="text-align:right;"> 11 </td>
#>    <td style="text-align:right;"> 13 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 3rd </td>
#>    <td style="text-align:right;"> 13 </td>
#>    <td style="text-align:right;"> 14 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> Crew </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>   </tr>
#>   <tr grouplength="4"><td colspan="3" style="border-bottom: 1px solid;"><strong>Survived: Yes / Age: Adult</strong></td></tr>
#> <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 1st </td>
#>    <td style="text-align:right;"> 57 </td>
#>    <td style="text-align:right;"> 140 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 2nd </td>
#>    <td style="text-align:right;"> 14 </td>
#>    <td style="text-align:right;"> 80 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 3rd </td>
#>    <td style="text-align:right;"> 75 </td>
#>    <td style="text-align:right;"> 76 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> Crew </td>
#>    <td style="text-align:right;"> 192 </td>
#>    <td style="text-align:right;"> 20 </td>
#>   </tr>
#> </tbody>
#> </table>
kable_array(Titanic, row_spec = list(0, angle = -30), 
  add_header_above =list(c(' '= 1,'Gender'=2,Freq = 1)),
  column_spec(1, bold = T, border_right = T)) 
#> [1] 1
#> Error in htmlTable_add_header_above(kable_input, header, bold, italic,     monospace, underline, strikeout, align, color, background,     font_size, angle, escape, line, line_sep, extra_css, include_empty): The new header row you provided has a total of 4 columns but the original kable_input has 3.
if (FALSE) { # \dontrun{  
kable_array(Titanic, row_spec = list(0, angle = -30), 
  add_header_above =list(c(' '= 1,'Gender'=2,Freq = 1)),
  column_spec(1, bold = T, border_right = T)) %>%
  save_kable(file = 'test.html', self_contained = T)
} # }