Effectiveness of migraine treatment

Subset of data on migraine treatments collected by Tammy Kostecki-Dillon. The data consists of headache logs kept by 133 patients in a treatment program in which bio-feedback was used to attempt to reduce migraine frequency and severity. Patients entered the program at different times over a period of about 3 years. Patients were encouraged to begin their logs four weeks before the onset of treatment and to continue for one month afterwards, but only 55 patients have data preceding the onset of treatment.

Usage

data(migraines)

Format

A data frame with 4152 observations on the following 9 variables.

id

patient id

ha

headache indictor: 1 if a headache has occurred on the day recorded, 0 otherwise

time

time in days relative to the onset of treatment which occurs at time == 0

dos

day of study in days from January 1 in the first year of the study

hatype

type of migraine headache experienced by a subject: Aura Mixed No Aura

female

an indicator variable

age

age at onset of treatment

airq

a measure of air quality

medication

medication status: Reduced Continuing None

Details

Source

References

Examples

data(migraines)
## maybe str(migraines) ; plot(migraines) ...