Overdispersion with glmer

Test for overdispersion and compute overdispersion factor for generalized linear mixed models.

Usage

overdisp_fun(model)

quasi_table(
  model,
  ctab = coef(summary(model)),
  phi = overdisp_fun(model)["ratio"]
)

Arguments

model

a model fitted using a generalized linear mixed model family without an independent scale parameter, i.e. poisson or binomial with counts (not with a 0/1 response for logistic regression)

Details

Two functions overdisp_fun and quasi_table are adapted from the from the GLMM FAQ by Ben Bolker.

Functions

quasi_table(): display coefficient table with standard errors adjusted for overdispersion

Examples

library(lme4)
#> Loading required package: Matrix
#> 
#> Attaching package: ‘lme4’
#> The following object is masked from ‘package:nlme’:
#> 
#>     lmList
#> The following object is masked from ‘package:spida2’:
#> 
#>     getData
library(spida2)
set.seed(123)
zd <- expand.grid(plate = 1:100) #treat = c('A','B'), batch = 1:10, plate = 1:5)
zd <- within(zd,
     {  
         treat <- plate %% 2 + 1
         batch <- plate %% 10 + 1
         eta <- rnorm(10)[batch]+ 0 * rnorm(100) + 5 * (treat == 2)
         eta_od <- rnorm(10)[batch]+ rnorm(100) + 5 * (treat == 2)
         count <- rpois(nrow(zd), exp(eta))
         count_od <- rpois(nrow(zd), exp(eta_od))
     }
)
fit <- glmer(count ~ treat + (1|batch), zd, family = 'poisson')
fit_od <- glmer(count_od ~ treat + (1|batch), zd, family = 'poisson')
summary(fit)
#> Generalized linear mixed model fit by maximum likelihood (Laplace
#>   Approximation) [glmerMod]
#>  Family: poisson  ( log )
#> Formula: count ~ treat + (1 | batch)
#>    Data: zd
#> 
#>      AIC      BIC   logLik deviance df.resid 
#>    585.5    593.3   -289.7    579.5       97 
#> 
#> Scaled residuals: 
#>     Min      1Q  Median      3Q     Max 
#> -2.1056 -0.7278 -0.2059  0.6172  3.1185 
#> 
#> Random effects:
#>  Groups Name        Variance Std.Dev.
#>  batch  (Intercept) 0.752    0.8672  
#> Number of obs: 100, groups:  batch, 10
#> 
#> Fixed effects:
#>             Estimate Std. Error z value Pr(>|z|)    
#> (Intercept)  -4.4820     0.9076  -4.938 7.87e-07 ***
#> treat         4.7201     0.5647   8.359  < 2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Correlation of Fixed Effects:
#>       (Intr)
#> treat -0.951
summary(fit_od)
#> Generalized linear mixed model fit by maximum likelihood (Laplace
#>   Approximation) [glmerMod]
#>  Family: poisson  ( log )
#> Formula: count_od ~ treat + (1 | batch)
#>    Data: zd
#> 
#>      AIC      BIC   logLik deviance df.resid 
#>  11310.6  11318.4  -5652.3  11304.6       97 
#> 
#> Scaled residuals: 
#>     Min      1Q  Median      3Q     Max 
#> -21.262  -4.062  -0.801   0.970  71.775 
#> 
#> Random effects:
#>  Groups Name        Variance Std.Dev.
#>  batch  (Intercept) 0.9675   0.9836  
#> Number of obs: 100, groups:  batch, 10
#> 
#> Fixed effects:
#>             Estimate Std. Error z value Pr(>|z|)    
#> (Intercept)  -5.3926     1.0480  -5.146 2.66e-07 ***
#> treat         5.3489     0.6479   8.256  < 2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Correlation of Fixed Effects:
#>       (Intr)
#> treat -0.951
overdisp_fun(fit)
#>      chisq      ratio        rdf          p 
#> 84.9286722  0.8755533 97.0000000  0.8042993 
overdisp_fun(fit_od)
#>      chisq      ratio        rdf          p 
#> 14866.0326   153.2581    97.0000     0.0000 
quasi_table(fit)
#>              Estimate Std. Error   z value     Pr(>|z|)
#> (Intercept) -4.482049  0.8492338 -5.277756 1.307754e-07
#> treat        4.720062  0.5283770  8.933132 4.141148e-19
quasi_table(fit_od)
#>              Estimate Std. Error    z value  Pr(>|z|)
#> (Intercept) -5.392646  12.973443 -0.4156680 0.6776529
#> treat        5.348902   8.020782  0.6668804 0.5048486
wald(fit)
#>  numDF denDF  F-value p-value
#>      2   Inf 81.78454 <.00001
#>             Estimate  Std.Error DF  t-value   p-value Lower 0.95 Upper 0.95
#> (Intercept) -4.482049 0.907582  Inf -4.938449 <.00001 -6.260878  -2.703221 
#> treat        4.720062 0.564680  Inf  8.358822 <.00001  3.613309   5.826815 

wald(fit_od)
#>  numDF denDF  F-value p-value
#>      2   Inf 72.68163 <.00001
#>             Estimate  Std.Error DF  t-value   p-value Lower 0.95 Upper 0.95
#> (Intercept) -5.392646 1.047957  Inf -5.145864 <.00001 -7.446604  -3.338687 
#> treat        5.348902 0.647896  Inf  8.255808 <.00001  4.079050   6.618754 
wald(fit_od, overdispersion = T)
#>  numDF denDF   F-value p-value overdispersion variance factor    chisq rdf p
#>      2   Inf 0.4742434 0.62236                       153.2581 14866.03  97 0
#>             Estimate  Std.Error DF  t-value   p-value Lower 0.95 Upper 0.95
#> (Intercept) -5.392646 12.973443 Inf -0.415668 0.67765 -30.82013  20.03484  
#> treat        5.348902  8.020782 Inf  0.666880 0.50485 -10.37154  21.06934