Draw data ellipse

This is an experiment in writing a function that can be called via layer or glayer without further complications e.g. xyplot(......,labels = rownames(data)) + layer(panel.dell(...)) or xyplot(....., labels = rownames(data), subscripts = T) + glayer(panel.dell(...)).

Usage

panel.dell(x, y, radius = 1, subscripts, type = "l", ...)

Arguments

x, y

points to generate data ellipse

radius

default 1

...

NOTE: may specify anything you don't want passed through ...

Examples

if (FALSE) { # \dontrun{
library(latticeExtra)
xyplot(mathach ~ ses, hs) + layer(panel.dell(..., col='red',lwd=2))
xyplot(mathach ~ ses | school, hs) + layer(panel.dell(...))
xyplot(mathach ~ ses , hs, groups = school) + glayer(panel.dell(...,lwd=2))
} # }