Construct pdInd object

This function is an alternative constructor for the pdMat class associated with object and is mostly used internally in other functions. See the documentation on the principal constructor function pdInd.

Usage

pdConstruct.pdInd(
  object,
  value = numeric(0),
  form = formula(object),
  nam = Names(object),
  data = sys.parent(),
  cov = NULL,
  ...
)

Arguments

object

an object inheriting from the class pdInd, representing a positive definite matrix with zero covariances except in the first row and column.

value

an optional initialization value

form

an optional one-sided linear formula specifying the row/column names for the matrix represented by object.

nam

and optional vector of character strings specifying the row/column names for the matrix represented by object.

data

and optional data frame i which to evaluate the variables names in value and form. ...

cov

optional position in lower triangle of covariances that are estimated and, thus, possibly non-zero. The default is that the covariances in the first column are estimated and possibly non-zero.