Hypothesis matrix for general regression splines

This function helps to build hypothesis matrices for general splines. See examples in gsp.

Usage

sc(sp, x, D = 0, type = 1)

Arguments

sp

a spline function generated by gsp

x

points where spline is tested

D

(default 0) order of derivative to test, 0 is the value of the spline, 1 is the first derivative, etc.

type

(default 1) if there is a discontinuity at a knot, type specifies whether to estimate the limit from the left (0), the limit from the right (1) or the saltus – limit from the right minus the limit from left (2)

See also