Substitute last occurrence of a pattern in a string

Substitute last occurrence of a pattern in a string

Usage

sublast(pattern, replacement, x, ...)

Arguments

pattern

character string containing a regular expression whose whose last match in 'x' will be replaced

replacement

character string that is a replacement for the matched 'pattern' in 'x'

x

a character vector where matches are sought.

...

arguments passed to gsub.

Examples

string <- 'a_b_c_D'
sublast('_','__',string)
#> [1] "a_b_c__D"
dd <- data.frame(id = 1:3, X_a_1 = 1:3, X_a_2 = 1:3, X_b_1 = 1:3, X_b_2 = 1:3)
dd
#>   id X_a_1 X_a_2 X_b_1 X_b_2
#> 1  1     1     1     1     1
#> 2  2     2     2     2     2
#> 3  3     3     3     3     3
names(dd) <- sublast('_','__',names(dd))
tolong(dd, sep = '__')
#>     id time X_a X_b
#> 1.1  1    1   1   1
#> 1.2  1    2   1   1
#> 2.1  2    1   2   2
#> 2.2  2    2   2   2
#> 3.1  3    1   3   3
#> 3.2  3    2   3   3
tolong(dd, sep = '__') %>% tolong(sep = '_', idvar = 'id2', timevar = 'time2')
#>     id time time2 X id2
#> 1.a  1    1     a 1   1
#> 1.b  1    1     b 1   1
#> 2.a  1    2     a 1   2
#> 2.b  1    2     b 1   2
#> 3.a  2    1     a 2   3
#> 3.b  2    1     b 2   3
#> 4.a  2    2     a 2   4
#> 4.b  2    2     b 2   4
#> 5.a  3    1     a 3   5
#> 5.b  3    1     b 3   5
#> 6.a  3    2     a 3   6
#> 6.b  3    2     b 3   6