Table of frequencies or relative frequencies bordered with totals and including NAs

Generates a table of frequencies or relative frequencies or relative percentages

Usage

tab(x, ...)

# S3 method for class 'table'
tab(x, ...)

# S3 method for class 'matrix'
tab(x, ...)

# S3 method for class 'array'
tab(x, ...)

# S3 method for class 'formula'
tab(fmla, data = sys.frame(sys.parent()), ...)

# S3 method for class 'data.frame'
tab(
  x,
  fmla,
  total.margins = TRUE,
  useNA = "ifany",
  pct = NULL,
  pr = NULL,
  test = FALSE,
  weights = NULL,
  na.rm = NULL,
  all.label = "All",
  simulate = FALSE,
  B = 2000,
  ...
)

# Default S3 method
tab(
  ...,
  total.margins = TRUE,
  pct = NULL,
  pr = NULL,
  useNA = "ifany",
  test = simulate,
  weights = NULL,
  na.rm = NULL,
  all.label = "All",
  simulate = FALSE,
  B = 2000
)

Arguments

...

as with table, one or more objects which can be interpreted as factors (including character strings), or a list (or data frame) whose components can be so interpreted.

fmla

a formula whose right-hand side names the variables to be used for tabulation. The optional left-hand side specifies a variable to be used for weights.

data

a data frame in which formula are interpreted

total.margins

if FALSE, generate table without margins

useNA

whether to include NA levels. The default is "ifany". Can also be set to "no" or "always".

pct

margins to be scaled to sum to 100. This is the vector of margin indices on which percentages are conditioned. For example, with the call tab(~ A + B + C, data, pct = 2:3), the table will contain conditional percentages for variable A conditional of combinations of B and C.

pr

margins to be scaled to sum to 1. This is the vector of margin indices on which percentages are conditioned.

test

(default FALSE) use chisq.test

weights

(not working temporarily) instead of generating a frequency table, generate a table with the sum of the weights

simulate

(default FALSE) simulate p-value with chisq.test

B

(default 2000) number of replications for simulation

keep

names of margins to keep with 'Tab', default = "All". To drop all margins, use default = "".

Value

An object of class 'table' of dimension equal to the number of variables, with optional margins showing totals. Elements of the matrix can be frequencies, relative frequencies, percentages or sums of weights.

Methods (by class)

tab(table): method class table

tab(matrix): method for matrices

tab(array): method for arrays

tab(formula): method for formulas

tab(data.frame): method for data frames

tab(default): default method

See also

Author

Georges Monette

Examples

titanic <- as.data.frame(Titanic)
head(titanic)
#>   Class    Sex   Age Survived Freq
#> 1   1st   Male Child       No    0
#> 2   2nd   Male Child       No    0
#> 3   3rd   Male Child       No   35
#> 4  Crew   Male Child       No    0
#> 5   1st Female Child       No    0
#> 6   2nd Female Child       No    0
tab(titanic, Freq ~ Sex + Survived + Age , test = T)
#> , , Age = Child
#> 
#>         Survived
#> Sex        No  Yes Total
#>   Male     35   29    64
#>   Female   17   28    45
#>   Total    52   57   109
#> 
#> , , Age = Adult
#> 
#>         Survived
#> Sex        No  Yes Total
#>   Male   1329  338  1667
#>   Female  109  316   425
#>   Total  1438  654  2092
#> 
#> , , Age = Total
#> 
#>         Survived
#> Sex        No  Yes Total
#>   Male   1364  367  1731
#>   Female  126  344   470
#>   Total  1490  711  2201
#> 
#> $Child
#> 
#> 	Pearson's Chi-squared test with Yates' continuity correction
#> 
#> data:  array(newX[, i], d.call, dn.call)
#> X-squared = 2.3885, df = 1, p-value = 0.1222
#> 
#> 
#> $Adult
#> 
#> 	Pearson's Chi-squared test with Yates' continuity correction
#> 
#> data:  array(newX[, i], d.call, dn.call)
#> X-squared = 458.35, df = 1, p-value < 2.2e-16
#> 
#> 
tab(titanic, Freq ~ Sex + Age)
#>         Age
#> Sex      Child Adult Total
#>   Male      64  1667  1731
#>   Female    45   425   470
#>   Total    109  2092  2201
tab(titanic, Freq ~ Sex + Survived + Age)
#> , , Age = Child
#> 
#>         Survived
#> Sex        No  Yes Total
#>   Male     35   29    64
#>   Female   17   28    45
#>   Total    52   57   109
#> 
#> , , Age = Adult
#> 
#>         Survived
#> Sex        No  Yes Total
#>   Male   1329  338  1667
#>   Female  109  316   425
#>   Total  1438  654  2092
#> 
#> , , Age = Total
#> 
#>         Survived
#> Sex        No  Yes Total
#>   Male   1364  367  1731
#>   Female  126  344   470
#>   Total  1490  711  2201
#> 
round(tab(titanic, Freq ~ Sex + Survived + Age,
    pct = c(1,3)),2)
#> , , Age = Child
#> 
#>         Survived
#> Sex          No    Yes  Total
#>   Male    54.69  45.31 100.00
#>   Female  37.78  62.22 100.00
#>   All     47.71  52.29 100.00
#> 
#> , , Age = Adult
#> 
#>         Survived
#> Sex          No    Yes  Total
#>   Male    79.72  20.28 100.00
#>   Female  25.65  74.35 100.00
#>   All     68.74  31.26 100.00
#> 
#> , , Age = All
#> 
#>         Survived
#> Sex          No    Yes  Total
#>   Male    78.80  21.20 100.00
#>   Female  26.81  73.19 100.00
#>   All     67.70  32.30 100.00
#> 
round(Tab(titanic, Freq ~ Sex + Survived + Age,
    pct = c(1,3)),2)
#> , , Age = Child
#> 
#>         Survived
#> Sex         No   Yes
#>   Male   54.69 45.31
#>   Female 37.78 62.22
#>   All    47.71 52.29
#> 
#> , , Age = Adult
#> 
#>         Survived
#> Sex         No   Yes
#>   Male   79.72 20.28
#>   Female 25.65 74.35
#>   All    68.74 31.26
#> 
#> , , Age = All
#> 
#>         Survived
#> Sex         No   Yes
#>   Male   78.80 21.20
#>   Female 26.81 73.19
#>   All    67.70 32.30
#> 
round(Tab(titanic, Freq ~ Sex + Survived + Age,
    pct = c(1,3), keep = ""),2)
#> , , Age = Child
#> 
#>         Survived
#> Sex         No   Yes
#>   Male   54.69 45.31
#>   Female 37.78 62.22
#>   All    47.71 52.29
#> 
#> , , Age = Adult
#> 
#>         Survived
#> Sex         No   Yes
#>   Male   79.72 20.28
#>   Female 25.65 74.35
#>   All    68.74 31.26
#> 
#> , , Age = All
#> 
#>         Survived
#> Sex         No   Yes
#>   Male   78.80 21.20
#>   Female 26.81 73.19
#>   All    67.70 32.30
#>