Create an array of data frames splitting on a formula with marginal faces

Like a table but each element consists of the rows of the data frame for a particular combination of levels of the formula

Usage

tabf(dd, fmla, all = TRUE, ...)

Arguments

dd

data frame

fmla

formula whose RHS creates a table and LHS adds variables to the resulting data frames

all

adds all variables in the data frame

Examples

tabf(mtcars, ~ cyl + carb)[1,1]            
#> [[1]]
#>                 mpg cyl  disp hp drat    wt  qsec vs am gear carb
#> Datsun 710     22.8   4 108.0 93 3.85 2.320 18.61  1  1    4    1
#> Fiat 128       32.4   4  78.7 66 4.08 2.200 19.47  1  1    4    1
#> Toyota Corolla 33.9   4  71.1 65 4.22 1.835 19.90  1  1    4    1
#> Toyota Corona  21.5   4 120.1 97 3.70 2.465 20.01  1  0    3    1
#> Fiat X1-9      27.3   4  79.0 66 4.08 1.935 18.90  1  1    4    1
#> 
tabf(mtcars, ~ cyl + carb) %>% dimnames  
#> [[1]]
#> [1] "1"   "2"   "3"   "All"
#> 
#> [[2]]
#> [1] "1"   "2"   "3"   "4"   "5"   "6"   "All"
#> 
tabf(mtcars, ~ cyl + carb + gear ) %>% Apply(dim) %>% Apply(paste, collapse = ' ')  
#> , , 1
#> 
#>     1      2      3      4      5      6      All    
#> 1   "1 11" "0 11" "0 11" "0 11" "0 11" "0 11" "1 11" 
#> 2   "2 11" "0 11" "0 11" "0 11" "0 11" "0 11" "2 11" 
#> 3   "0 11" "4 11" "3 11" "5 11" "0 11" "0 11" "12 11"
#> All "3 11" "4 11" "3 11" "5 11" "0 11" "0 11" "15 11"
#> 
#> , , 2
#> 
#>     1      2      3      4      5      6      All    
#> 1   "4 11" "4 11" "0 11" "0 11" "0 11" "0 11" "8 11" 
#> 2   "0 11" "0 11" "0 11" "4 11" "0 11" "0 11" "4 11" 
#> 3   "0 11" "0 11" "0 11" "0 11" "0 11" "0 11" "0 11" 
#> All "4 11" "4 11" "0 11" "4 11" "0 11" "0 11" "12 11"
#> 
#> , , 3
#> 
#>     1      2      3      4      5      6      All   
#> 1   "0 11" "2 11" "0 11" "0 11" "0 11" "0 11" "2 11"
#> 2   "0 11" "0 11" "0 11" "0 11" "1 11" "0 11" "1 11"
#> 3   "0 11" "0 11" "0 11" "1 11" "0 11" "1 11" "2 11"
#> All "0 11" "2 11" "0 11" "1 11" "1 11" "1 11" "5 11"
#> 
#> , , All
#> 
#>     1      2       3      4       5      6      All    
#> 1   "5 11" "6 11"  "0 11" "0 11"  "0 11" "0 11" "11 11"
#> 2   "2 11" "0 11"  "0 11" "4 11"  "1 11" "0 11" "7 11" 
#> 3   "0 11" "4 11"  "3 11" "6 11"  "0 11" "1 11" "14 11"
#> All "7 11" "10 11" "3 11" "10 11" "1 11" "1 11" "32 11"
#>