Create a long file from a wide file

Uses a minimal number of arguments to create a long file using stats::reshape. Produces output even when long variable names and time values are not fully crossed.

Usage

tolong(
  data,
  sep = "_",
  timevar = "time",
  idvar = "id",
  ids = 1:nrow(data),
  valuepattern = if (numericalpattern) "[0-9]+" else ".*",
  numericalpattern = FALSE,
  expand = TRUE,
  safe_sep = "@3-2861-2579@",
  reverse = F,
  ...
)

tolong_old(
  data,
  sep = "_",
  timevar = "time",
  idvar = "id",
  ids = 1:nrow(data),
  expand = TRUE,
  safe_sep = "#%@!",
  reverse = F,
  ...
)

long(
  data,
  sep = "_",
  timevar = "time",
  idvar = "id",
  ids = 1:nrow(data),
  valuepattern = if (numericalpattern) "[0-9]+" else ".*",
  numericalpattern = FALSE,
  expand = TRUE,
  safe_sep = "@3-2861-2579@",
  reverse = F,
  ...
)

Arguments

data

wide data frame

sep

(default '_') single or multiple character separator between long names and 'time' value. Variable names with this separator are transformed to long variables. If the string occurs multiple times in a variable name, only the last occurrence is treated as a separator.

timevar

(default 'time') names of variable in output long file to identify occasions. Its values are taken from the suffix following the 'sep' character in each time-varying variable.

idvar

(default: 'id') the variable name used in the output long file to identify the provenance row in the input wide file. It may exist in the input wide file and must, in that case, have a unique value in each row. If it does not exist, it will be created with values equal to the row numbers of the input wide file.

ids

(default 1:nrow(data)) values for idvar in long file if the variable idvar does not exist in the input wide file. Ignored if idvar exists in data.

valuepattern

a regular expression to match the form of time values at the end of variable names immediately following the separator. Specifying this pattern can avoid misinterpreting separators in variable names that are not intended to turned into long variables.

numericalpattern

(default FALSE) if TRUE, valuepattern is set to "0-9+$" to match numerical values as time values at the end of variable names to be turned into long variables. This parameter is only necessary if the separator occurs in variable names where it is not intended to identify a long variable.

expand

(default TRUE): if 'time' values are inconsistent, fill in missing 'time's with NAs.

safe_sep

temporary safe? separator

reverse

(default FALSE) if TRUE, the 'time' value precedes the variable name

...

additional parameters are passed to reshape.

Value

'long' file with each wide row repeated as many times as there are distinct values for the 'timevar' variable. The rownames show the provenance of each row by combining the value of id with the value of time separated by a period

Details

tolong is intended for the simple case in which 'wide' variables in the input data frame are separator string that separates the name of the variable in the long file from the value of the 'time' variable that identifies the corresponding row in the long file, e.g x_1, x_2, x_3 or brain.volume_left, brain.volume_right. Since the separator ('_' by default) may occur in other variables, tolong offers two mechanisms to avoid misinterpreting those occurrences as separators. If there are multiple occurrences of the separator string in a variable name, only the last occurrence is interpreted as a separator. Secondly, the valuepattern parameter can specify a regular expression to identify allowed 'time' value strings at the end of variable names. The common case where the 'time' values are numerical can be specified with the numericalpattern parameter.

reshape does not work if long variable names and time values are not fully crossed, e.g x_1, x_2, x_3, y_1, y_2. By default long creates additional variables with "NAs" so the set of variables given to reshape is fully crossed, e.g. adding a variable y_3 <- NA.

Compare the functionality of tolong with that of tidyr::gather and of tidyr::pivot_longer. 'tolong' depends on the format of variable names to identify variables whose values become new variables in the long form of the data and which labels are used as the indices of the indexing variable, whose default name is 'time', which can be set to another value with the "timevar" argument. "tolong" can handle many 'time-varying' variables. "gather" can only handle one. "pivot_longer" can handle many and might be considered a replacement for "to_long" which has the disadvantage of frequently requiring the renaming of variables, an easier task for those who have mastered the use of regular expressions, but potentially challenging otherwise.

See also

Examples

z <- data.frame(id =letters[1:10], id2= 11:20, v_L = 1:10, v_R = 11:20)
z
#>    id id2 v_L v_R
#> 1   a  11   1  11
#> 2   b  12   2  12
#> 3   c  13   3  13
#> 4   d  14   4  14
#> 5   e  15   5  15
#> 6   f  16   6  16
#> 7   g  17   7  17
#> 8   h  18   8  18
#> 9   i  19   9  19
#> 10  j  20  10  20
tolong(z)
#>     id id2 time  v
#> a.L  a  11    L  1
#> a.R  a  11    R 11
#> b.L  b  12    L  2
#> b.R  b  12    R 12
#> c.L  c  13    L  3
#> c.R  c  13    R 13
#> d.L  d  14    L  4
#> d.R  d  14    R 14
#> e.L  e  15    L  5
#> e.R  e  15    R 15
#> f.L  f  16    L  6
#> f.R  f  16    R 16
#> g.L  g  17    L  7
#> g.R  g  17    R 17
#> h.L  h  18    L  8
#> h.R  h  18    R 18
#> i.L  i  19    L  9
#> i.R  i  19    R 19
#> j.L  j  20    L 10
#> j.R  j  20    R 20
tolong(z, timevar = 'Side', idvar = 'idn', ids = LETTERS[1:10])
#>     id id2 Side  v idn
#> A.L  a  11    L  1   A
#> A.R  a  11    R 11   A
#> B.L  b  12    L  2   B
#> B.R  b  12    R 12   B
#> C.L  c  13    L  3   C
#> C.R  c  13    R 13   C
#> D.L  d  14    L  4   D
#> D.R  d  14    R 14   D
#> E.L  e  15    L  5   E
#> E.R  e  15    R 15   E
#> F.L  f  16    L  6   F
#> F.R  f  16    R 16   F
#> G.L  g  17    L  7   G
#> G.R  g  17    R 17   G
#> H.L  h  18    L  8   H
#> H.R  h  18    R 18   H
#> I.L  i  19    L  9   I
#> I.R  i  19    R 19   I
#> J.L  j  20    L 10   J
#> J.R  j  20    R 20   J
tolong(z, timevar = 'Side', idvar = 'idn', ids = z$id2)
#>      id id2 Side  v idn
#> 11.L  a  11    L  1  11
#> 11.R  a  11    R 11  11
#> 12.L  b  12    L  2  12
#> 12.R  b  12    R 12  12
#> 13.L  c  13    L  3  13
#> 13.R  c  13    R 13  13
#> 14.L  d  14    L  4  14
#> 14.R  d  14    R 14  14
#> 15.L  e  15    L  5  15
#> 15.R  e  15    R 15  15
#> 16.L  f  16    L  6  16
#> 16.R  f  16    R 16  16
#> 17.L  g  17    L  7  17
#> 17.R  g  17    R 17  17
#> 18.L  h  18    L  8  18
#> 18.R  h  18    R 18  18
#> 19.L  i  19    L  9  19
#> 19.R  i  19    R 19  19
#> 20.L  j  20    L 10  20
#> 20.R  j  20    R 20  20

# unbalanced times
z <- data.frame(id =letters[1:10], id2= 11:20, v_L = 1:10, v_R = 11:20, z_L = 21:30)
z
#>    id id2 v_L v_R z_L
#> 1   a  11   1  11  21
#> 2   b  12   2  12  22
#> 3   c  13   3  13  23
#> 4   d  14   4  14  24
#> 5   e  15   5  15  25
#> 6   f  16   6  16  26
#> 7   g  17   7  17  27
#> 8   h  18   8  18  28
#> 9   i  19   9  19  29
#> 10  j  20  10  20  30
tolong(z)
#>     id id2 time  v  z
#> a.L  a  11    L  1 21
#> a.R  a  11    R 11 NA
#> b.L  b  12    L  2 22
#> b.R  b  12    R 12 NA
#> c.L  c  13    L  3 23
#> c.R  c  13    R 13 NA
#> d.L  d  14    L  4 24
#> d.R  d  14    R 14 NA
#> e.L  e  15    L  5 25
#> e.R  e  15    R 15 NA
#> f.L  f  16    L  6 26
#> f.R  f  16    R 16 NA
#> g.L  g  17    L  7 27
#> g.R  g  17    R 17 NA
#> h.L  h  18    L  8 28
#> h.R  h  18    R 18 NA
#> i.L  i  19    L  9 29
#> i.R  i  19    R 19 NA
#> j.L  j  20    L 10 30
#> j.R  j  20    R 20 NA

# a separator with multiple occurrences:
z <- data.frame(id =letters[1:10], id2= 11:20, v_a_L = 1:10, v_a_R = 11:20, z_L = 21:30)
z
#>    id id2 v_a_L v_a_R z_L
#> 1   a  11     1    11  21
#> 2   b  12     2    12  22
#> 3   c  13     3    13  23
#> 4   d  14     4    14  24
#> 5   e  15     5    15  25
#> 6   f  16     6    16  26
#> 7   g  17     7    17  27
#> 8   h  18     8    18  28
#> 9   i  19     9    19  29
#> 10  j  20    10    20  30
# The previous version of tolong() would have produced an error 
# due to multiple occurrences of the default separator '_'
# but the new version matches only the last occurrence in
# each variable name. The sublast() function helps by
# matching only the last occurrence to facilitate 
# replacing it with a new unique separator, but it is
# no longer necessary to do this:
zz <- z
names(zz) <- sublast('_', '__', names(zz))
tolong(zz, sep = '__')
#>     id id2 time v_a  z
#> a.L  a  11    L   1 21
#> a.R  a  11    R  11 NA
#> b.L  b  12    L   2 22
#> b.R  b  12    R  12 NA
#> c.L  c  13    L   3 23
#> c.R  c  13    R  13 NA
#> d.L  d  14    L   4 24
#> d.R  d  14    R  14 NA
#> e.L  e  15    L   5 25
#> e.R  e  15    R  15 NA
#> f.L  f  16    L   6 26
#> f.R  f  16    R  16 NA
#> g.L  g  17    L   7 27
#> g.R  g  17    R  17 NA
#> h.L  h  18    L   8 28
#> h.R  h  18    R  18 NA
#> i.L  i  19    L   9 29
#> i.R  i  19    R  19 NA
#> j.L  j  20    L  10 30
#> j.R  j  20    R  20 NA
# or, now,, with the same result:
tolong(z)
#>     id id2 time v_a  z
#> a.L  a  11    L   1 21
#> a.R  a  11    R  11 NA
#> b.L  b  12    L   2 22
#> b.R  b  12    R  12 NA
#> c.L  c  13    L   3 23
#> c.R  c  13    R  13 NA
#> d.L  d  14    L   4 24
#> d.R  d  14    R  14 NA
#> e.L  e  15    L   5 25
#> e.R  e  15    R  15 NA
#> f.L  f  16    L   6 26
#> f.R  f  16    R  16 NA
#> g.L  g  17    L   7 27
#> g.R  g  17    R  17 NA
#> h.L  h  18    L   8 28
#> h.R  h  18    R  18 NA
#> i.L  i  19    L   9 29
#> i.R  i  19    R  19 NA
#> j.L  j  20    L  10 30
#> j.R  j  20    R  20 NA
#
# - sep can use more than one character
# - the character string is interpreted literally, 
# i.e. if special regular expression characters
# they are interpreted literally.
z <- data.frame(id =letters[1:10], id2= 11:20, HPC_head_R = 1:10, HPC_tail_R = 11:20, HPC_head_L = 21:30, HPC_tail_L = 31:40)
z
#>    id id2 HPC_head_R HPC_tail_R HPC_head_L HPC_tail_L
#> 1   a  11          1         11         21         31
#> 2   b  12          2         12         22         32
#> 3   c  13          3         13         23         33
#> 4   d  14          4         14         24         34
#> 5   e  15          5         15         25         35
#> 6   f  16          6         16         26         36
#> 7   g  17          7         17         27         37
#> 8   h  18          8         18         28         38
#> 9   i  19          9         19         29         39
#> 10  j  20         10         20         30         40
names(z) <- sub("(_[LR]$)","_\\1", names(z))
names(z)
#> [1] "id"          "id2"         "HPC_head__R" "HPC_tail__R" "HPC_head__L"
#> [6] "HPC_tail__L"
(zz <- tolong(z, sep = "__", timevar = "Side"))
#>     id id2 Side HPC_head HPC_tail
#> a.R  a  11    R        1       11
#> a.L  a  11    L       21       31
#> b.R  b  12    R        2       12
#> b.L  b  12    L       22       32
#> c.R  c  13    R        3       13
#> c.L  c  13    L       23       33
#> d.R  d  14    R        4       14
#> d.L  d  14    L       24       34
#> e.R  e  15    R        5       15
#> e.L  e  15    L       25       35
#> f.R  f  16    R        6       16
#> f.L  f  16    L       26       36
#> g.R  g  17    R        7       17
#> g.L  g  17    L       27       37
#> h.R  h  18    R        8       18
#> h.L  h  18    L       28       38
#> i.R  i  19    R        9       19
#> i.L  i  19    L       29       39
#> j.R  j  20    R       10       20
#> j.L  j  20    L       30       40
zz$id3 <- rownames(zz)
tolong(zz, idvar = 'id3' ,timevar = 'Part')
#>          id id2 Side id3 Part HPC
#> a.L.head  a  11    L a.L head  21
#> a.L.tail  a  11    L a.L tail  31
#> a.R.head  a  11    R a.R head   1
#> a.R.tail  a  11    R a.R tail  11
#> b.L.head  b  12    L b.L head  22
#> b.L.tail  b  12    L b.L tail  32
#> b.R.head  b  12    R b.R head   2
#> b.R.tail  b  12    R b.R tail  12
#> c.L.head  c  13    L c.L head  23
#> c.L.tail  c  13    L c.L tail  33
#> c.R.head  c  13    R c.R head   3
#> c.R.tail  c  13    R c.R tail  13
#> d.L.head  d  14    L d.L head  24
#> d.L.tail  d  14    L d.L tail  34
#> d.R.head  d  14    R d.R head   4
#> d.R.tail  d  14    R d.R tail  14
#> e.L.head  e  15    L e.L head  25
#> e.L.tail  e  15    L e.L tail  35
#> e.R.head  e  15    R e.R head   5
#> e.R.tail  e  15    R e.R tail  15
#> f.L.head  f  16    L f.L head  26
#> f.L.tail  f  16    L f.L tail  36
#> f.R.head  f  16    R f.R head   6
#> f.R.tail  f  16    R f.R tail  16
#> g.L.head  g  17    L g.L head  27
#> g.L.tail  g  17    L g.L tail  37
#> g.R.head  g  17    R g.R head   7
#> g.R.tail  g  17    R g.R tail  17
#> h.L.head  h  18    L h.L head  28
#> h.L.tail  h  18    L h.L tail  38
#> h.R.head  h  18    R h.R head   8
#> h.R.tail  h  18    R h.R tail  18
#> i.L.head  i  19    L i.L head  29
#> i.L.tail  i  19    L i.L tail  39
#> i.R.head  i  19    R i.R head   9
#> i.R.tail  i  19    R i.R tail  19
#> j.L.head  j  20    L j.L head  30
#> j.L.tail  j  20    L j.L tail  40
#> j.R.head  j  20    R j.R head  10
#> j.R.tail  j  20    R j.R tail  20

dd <- data.frame( y.a = 1:3, y.b = 1:3, x.a= 1:3, time = 1:3,
    x.b = 11:13, x.c = 21:23, id = c('a','a','b'))
dd
#>   y.a y.b x.a time x.b x.c id
#> 1   1   1   1    1  11  21  a
#> 2   2   2   2    2  12  22  a
#> 3   3   3   3    3  13  23  b
tolong(dd, sep = '.')
#> Warning: Variable time in data is replaced by a variable to mark occasions. Use the 'timevar' argument to specify a different variable name in the output data frame
#> Warning: idvar: id must have unique values. It will be replaced in the output data frame with a variable containing row numbers of the input data frame. Use the 'idvar' argument to specify a different variable name in the output data frame
#>     time id  y  x
#> 1.a    a  1  1  1
#> 1.b    b  1  1 11
#> 1.c    c  1 NA 21
#> 2.a    a  2  2  2
#> 2.b    b  2  2 12
#> 2.c    c  2 NA 22
#> 3.a    a  3  3  3
#> 3.b    b  3  3 13
#> 3.c    c  3 NA 23
dl <- tolong(dd, sep = '.', timevar = "type", idvar = 'patient')
dl
#>     time id type  y  x patient
#> 1.a    1  a    a  1  1       1
#> 1.b    1  a    b  1 11       1
#> 1.c    1  a    c NA 21       1
#> 2.a    2  a    a  2  2       2
#> 2.b    2  a    b  2 12       2
#> 2.c    2  a    c NA 22       2
#> 3.a    3  b    a  3  3       3
#> 3.b    3  b    b  3 13       3
#> 3.c    3  b    c NA 23       3
towide(dl, idvar = 'patient', timevar = 'type')
#>   patient y_a x_a y_b x_b y_c x_c time id
#> 1       1   1   1   1  11  NA  21    1  a
#> 2       2   2   2   2  12  NA  22    2  a
#> 3       3   3   3   3  13  NA  23    3  b

# Long file with additional constants

dl <- data.frame(name = rep(c('A','B','C'), c(3,3,2)),
                 site = c('head','neck','jaw','chest')[
                   c(1,2,3,1,2,3,1,4)],
                 sex = rep(c('male','female','male'), c(3,3,2)),
                 var1 = 1:8,
                 var2 = 11:18,
                 invar = rep(1:3, c(3,3,2)))
towide(dl, c('name'), 'site')
#>   name var1_head var2_head var1_neck var2_neck var1_jaw var2_jaw var1_chest
#> 1    A         1        11         2        12        3       13         NA
#> 2    B         4        14         5        15        6       16         NA
#> 3    C         7        17        NA        NA       NA       NA          8
#>   var2_chest    sex invar
#> 1         NA   male     1
#> 2         NA female     2
#> 3         18   male     3
#
# Two indexing variable: e.g. hippocampal volume 2 sides x 3 sites
#
dl <- data.frame(name = rep(LETTERS[1:3], each = 6),
                 side = rep(c('left','right'), 9),
                 site = rep(rep(c('head','body','tail'),each = 2),3),
                 volume = 1:18,
                 sex = rep(c('female','male','female'), each = 6),
                 age = rep(c(25, 43, 69), each = 6))
dl
#>    name  side site volume    sex age
#> 1     A  left head      1 female  25
#> 2     A right head      2 female  25
#> 3     A  left body      3 female  25
#> 4     A right body      4 female  25
#> 5     A  left tail      5 female  25
#> 6     A right tail      6 female  25
#> 7     B  left head      7   male  43
#> 8     B right head      8   male  43
#> 9     B  left body      9   male  43
#> 10    B right body     10   male  43
#> 11    B  left tail     11   male  43
#> 12    B right tail     12   male  43
#> 13    C  left head     13 female  69
#> 14    C right head     14 female  69
#> 15    C  left body     15 female  69
#> 16    C right body     16 female  69
#> 17    C  left tail     17 female  69
#> 18    C right tail     18 female  69
(dlsite <- towide(dl, c('name','side'), 'site'))
#>   name  side volume_head volume_body volume_tail    sex age
#> 1    A  left           1           3           5 female  25
#> 2    A right           2           4           6 female  25
#> 3    B  left           7           9          11   male  43
#> 4    B right           8          10          12   male  43
#> 5    C  left          13          15          17 female  69
#> 6    C right          14          16          18 female  69
(dlsite.side <- towide(dlsite, c('name'), 'side'))
#>   name volume_head_left volume_body_left volume_tail_left volume_head_right
#> 1    A                1                3                5                 2
#> 2    B                7                9               11                 8
#> 3    C               13               15               17                14
#>   volume_body_right volume_tail_right    sex age
#> 1                 4                 6 female  25
#> 2                10                12   male  43
#> 3                16                18 female  69
#
# Flipping a data frame
#
z <- data.frame(vname = 
   rep(c('v1','v2','v3'), each = 4),
   country = rep(c('Angola','Benin','Chad','Denmark'), 3),
   code = rep(c('ANG','BEN','CHA','DEN'),3),
   val__2011 = 2011 + seq(.01,.12,.01),
   val__2012 = 2012 + seq(.01,.12,.01),
   val__2013 = 2013 + seq(.01,.12,.01),
   val__2014 = 2014 + seq(.01,.12,.01),
   val__2015 = 2015 + seq(.01,.12,.01)
)
z
#>    vname country code val__2011 val__2012 val__2013 val__2014 val__2015
#> 1     v1  Angola  ANG   2011.01   2012.01   2013.01   2014.01   2015.01
#> 2     v1   Benin  BEN   2011.02   2012.02   2013.02   2014.02   2015.02
#> 3     v1    Chad  CHA   2011.03   2012.03   2013.03   2014.03   2015.03
#> 4     v1 Denmark  DEN   2011.04   2012.04   2013.04   2014.04   2015.04
#> 5     v2  Angola  ANG   2011.05   2012.05   2013.05   2014.05   2015.05
#> 6     v2   Benin  BEN   2011.06   2012.06   2013.06   2014.06   2015.06
#> 7     v2    Chad  CHA   2011.07   2012.07   2013.07   2014.07   2015.07
#> 8     v2 Denmark  DEN   2011.08   2012.08   2013.08   2014.08   2015.08
#> 9     v3  Angola  ANG   2011.09   2012.09   2013.09   2014.09   2015.09
#> 10    v3   Benin  BEN   2011.10   2012.10   2013.10   2014.10   2015.10
#> 11    v3    Chad  CHA   2011.11   2012.11   2013.11   2014.11   2015.11
#> 12    v3 Denmark  DEN   2011.12   2012.12   2013.12   2014.12   2015.12
z %>% 
  tolong(sep= '__') 
#>         vname country code time     val id
#> 1.2011     v1  Angola  ANG 2011 2011.01  1
#> 1.2012     v1  Angola  ANG 2012 2012.01  1
#> 1.2013     v1  Angola  ANG 2013 2013.01  1
#> 1.2014     v1  Angola  ANG 2014 2014.01  1
#> 1.2015     v1  Angola  ANG 2015 2015.01  1
#> 2.2011     v1   Benin  BEN 2011 2011.02  2
#> 2.2012     v1   Benin  BEN 2012 2012.02  2
#> 2.2013     v1   Benin  BEN 2013 2013.02  2
#> 2.2014     v1   Benin  BEN 2014 2014.02  2
#> 2.2015     v1   Benin  BEN 2015 2015.02  2
#> 3.2011     v1    Chad  CHA 2011 2011.03  3
#> 3.2012     v1    Chad  CHA 2012 2012.03  3
#> 3.2013     v1    Chad  CHA 2013 2013.03  3
#> 3.2014     v1    Chad  CHA 2014 2014.03  3
#> 3.2015     v1    Chad  CHA 2015 2015.03  3
#> 4.2011     v1 Denmark  DEN 2011 2011.04  4
#> 4.2012     v1 Denmark  DEN 2012 2012.04  4
#> 4.2013     v1 Denmark  DEN 2013 2013.04  4
#> 4.2014     v1 Denmark  DEN 2014 2014.04  4
#> 4.2015     v1 Denmark  DEN 2015 2015.04  4
#> 5.2011     v2  Angola  ANG 2011 2011.05  5
#> 5.2012     v2  Angola  ANG 2012 2012.05  5
#> 5.2013     v2  Angola  ANG 2013 2013.05  5
#> 5.2014     v2  Angola  ANG 2014 2014.05  5
#> 5.2015     v2  Angola  ANG 2015 2015.05  5
#> 6.2011     v2   Benin  BEN 2011 2011.06  6
#> 6.2012     v2   Benin  BEN 2012 2012.06  6
#> 6.2013     v2   Benin  BEN 2013 2013.06  6
#> 6.2014     v2   Benin  BEN 2014 2014.06  6
#> 6.2015     v2   Benin  BEN 2015 2015.06  6
#> 7.2011     v2    Chad  CHA 2011 2011.07  7
#> 7.2012     v2    Chad  CHA 2012 2012.07  7
#> 7.2013     v2    Chad  CHA 2013 2013.07  7
#> 7.2014     v2    Chad  CHA 2014 2014.07  7
#> 7.2015     v2    Chad  CHA 2015 2015.07  7
#> 8.2011     v2 Denmark  DEN 2011 2011.08  8
#> 8.2012     v2 Denmark  DEN 2012 2012.08  8
#> 8.2013     v2 Denmark  DEN 2013 2013.08  8
#> 8.2014     v2 Denmark  DEN 2014 2014.08  8
#> 8.2015     v2 Denmark  DEN 2015 2015.08  8
#> 9.2011     v3  Angola  ANG 2011 2011.09  9
#> 9.2012     v3  Angola  ANG 2012 2012.09  9
#> 9.2013     v3  Angola  ANG 2013 2013.09  9
#> 9.2014     v3  Angola  ANG 2014 2014.09  9
#> 9.2015     v3  Angola  ANG 2015 2015.09  9
#> 10.2011    v3   Benin  BEN 2011 2011.10 10
#> 10.2012    v3   Benin  BEN 2012 2012.10 10
#> 10.2013    v3   Benin  BEN 2013 2013.10 10
#> 10.2014    v3   Benin  BEN 2014 2014.10 10
#> 10.2015    v3   Benin  BEN 2015 2015.10 10
#> 11.2011    v3    Chad  CHA 2011 2011.11 11
#> 11.2012    v3    Chad  CHA 2012 2012.11 11
#> 11.2013    v3    Chad  CHA 2013 2013.11 11
#> 11.2014    v3    Chad  CHA 2014 2014.11 11
#> 11.2015    v3    Chad  CHA 2015 2015.11 11
#> 12.2011    v3 Denmark  DEN 2011 2011.12 12
#> 12.2012    v3 Denmark  DEN 2012 2012.12 12
#> 12.2013    v3 Denmark  DEN 2013 2013.12 12
#> 12.2014    v3 Denmark  DEN 2014 2014.12 12
#> 12.2015    v3 Denmark  DEN 2015 2015.12 12

z %>% 
  tolong(sep= '__', timevar = 'year') %>% 
  .[!grepl('^id$',names(.))] %>% 
  towide(timevar = 'vname', idvar = c('code','year'))
#>    code year  val_v1  val_v2  val_v3 country
#> 1   ANG 2011 2011.01 2011.05 2011.09  Angola
#> 2   ANG 2012 2012.01 2012.05 2012.09  Angola
#> 3   ANG 2013 2013.01 2013.05 2013.09  Angola
#> 4   ANG 2014 2014.01 2014.05 2014.09  Angola
#> 5   ANG 2015 2015.01 2015.05 2015.09  Angola
#> 6   BEN 2011 2011.02 2011.06 2011.10   Benin
#> 7   BEN 2012 2012.02 2012.06 2012.10   Benin
#> 8   BEN 2013 2013.02 2013.06 2013.10   Benin
#> 9   BEN 2014 2014.02 2014.06 2014.10   Benin
#> 10  BEN 2015 2015.02 2015.06 2015.10   Benin
#> 11  CHA 2011 2011.03 2011.07 2011.11    Chad
#> 12  CHA 2012 2012.03 2012.07 2012.11    Chad
#> 13  CHA 2013 2013.03 2013.07 2013.11    Chad
#> 14  CHA 2014 2014.03 2014.07 2014.11    Chad
#> 15  CHA 2015 2015.03 2015.07 2015.11    Chad
#> 16  DEN 2011 2011.04 2011.08 2011.12 Denmark
#> 17  DEN 2012 2012.04 2012.08 2012.12 Denmark
#> 18  DEN 2013 2013.04 2013.08 2013.12 Denmark
#> 19  DEN 2014 2014.04 2014.08 2014.12 Denmark
#> 20  DEN 2015 2015.04 2015.08 2015.12 Denmark

if (FALSE) { # \dontrun{
# Extracting chains from a stanfit object in the 'rstan' package
# If 'mod' is a stanfit model
library(rstan)
library(spida2)
df <- as.data.frame(extract(mod, permute = F))
dl <- tolong(df, sep = ':', reverse = T)
} # }