Create a data frame at a higher level of aggregation

Produce a higher level data set with one row per cluster. The data set can contain only variables that are invariant in each cluster or it can also include summaries (mean or modes) of variables that vary by cluster. Adapted from gsummary in the nlme package.

Usage

up(
  object,
  form = formula(object),
  agg = NULL,
  sep.agg = "_",
  freq = NULL,
  sep.freq = "_",
  all = FALSE,
  sep = "/",
  na.rm = TRUE,
  FUN = function(x) mean(x, na.rm = na.rm),
  omitGroupingFactor = FALSE,
  groups,
  invariantsOnly = !all,
  ...
)

Arguments

object

a data frame to be aggregated.

form

a one-sided formula identifying the variable(s) in object that identifies clusters. e.g. ~ school/Sex to get a summary within each Sex of each school.

agg

(NEW: Aug 2016) a one-sided formula identifying variables to be aggregated, i.e. variables that vary within clusters and that need to be aggregated (within-cluster mean for numeric variables and within-cluster incidence proportions for factors). Default: NULL

sep.agg

(NEW: Aug 2016) separator between factor names and factor level for within-cluster incidence proportions. Default: '_'

freq

(NEW: Nov 2018) a one-sided formula identifying character variables to be represented according the frequencies of their levels, i.e. variables that vary withing cluster and that need to be aggregated (within-cluster sum for numeric variables and within-cluster frequencies for factors). Default: NULL

sep.freq

(NEW: Nove 2018) separator between factor names and factor level for within-cluster incidence frequencies. Default: '_'

all

if TRUE, include summaries of variables that vary within clusters, otherwise keep only cluster-invariant variables and variables listed in 'agg'

sep

separator to form cluster names combining more than one clustering variables. If the separator leads to the same name for distinct clusters (e.g. if var1 has levels 'a' and 'a/b' and var2 has levels 'b/c' and 'c') the function produces an error and a different separator should be used.

FUN

function to be used for summaries.

omitGroupingFactor

kept for compatibility with gsummary

groups

kept for compatibility with gsummary

invariantsOnly

kept for compatibility with gsummary

...

additional arguments to tapply when summarizing numerical variables. e.g. na.rm = TRUE

Value

a data frame with one row per combination of values of the variable(s) in form. The number of rows for each combination is retuned in a variable 'Freq'. Frequencies (proportions) of values for each variable specified by '~freq' ('~agg') are also included.

Details

up was created from nlme::gsummary and modified to make it easier to use and to make an equivalent of gsummary available when using lme4.

Author

adapted by G. Monette from gsummary in 'nlme' by Bates & Pinheiro

Examples

    data(hs)
    dim( hs )
#> [1] 1977    9
    hsu <- up( hs, ~ school )
    dim( hsu )
#> [1] 40  6

    # to also get cluster means of cluster-varying numeric variables and modes of factors:

    hsa <- up( hs, ~ school , all = TRUE )

    # to get summary proportions of cluster varying factors:

    up( cbind( hs, model.matrix( ~ Sex -1 , hs)), ~ school, all = T)
#>      school   mathach         ses    Sex Minority Size   Sector PRACAD DISCLIM
#> 1317   1317 13.177687  0.34533333 Female      Yes  455 Catholic   0.95  -1.694
#> 1906   1906 15.983170  0.51162264 Female       No  400 Catholic   0.87  -0.939
#> 2208   2208 15.404667  0.42316667 Female       No 1061 Catholic   0.68  -0.864
#> 2458   2458 13.985684  0.22778947 Female      Yes  545 Catholic   0.89  -1.484
#> 2626   2626 13.396605 -0.06484211   Male       No 2142   Public   0.40   0.142
#> 2629   2629 14.907772 -0.13764912   Male       No 1314 Catholic   0.81  -0.613
#> 2639   2639  6.615476 -0.96442857 Female      Yes 2713   Public   0.14  -0.282
#> 2658   2658 13.396156  0.43844444 Female       No  780 Catholic   0.79  -0.961
#> 2771   2771 11.844109 -0.33945455 Female       No  415   Public   0.24   1.048
#> 3013   3013 12.610830 -0.04422642   Male       No  760   Public   0.56  -0.213
#> 3610   3610 15.354953  0.12012500   Male       No 1431 Catholic   0.80  -0.621
#> 3992   3992 14.645208  0.36539623   Male       No 1114 Catholic   0.73  -1.534
#> 4292   4292 12.864354 -0.48615385   Male      Yes 1328 Catholic   0.76  -0.674
#> 4511   4511 13.409034 -0.10713793 Female       No 1068 Catholic   0.52  -1.872
#> 4530   4530  9.055698 -0.59688889 Female       No  435 Catholic   0.60  -0.245
#> 4868   4868 12.310176  0.36111765   Male       No  657 Catholic   1.00  -0.219
#> 5619   5619 15.416242  0.42033333   Male       No 1118 Catholic   0.77  -1.286
#> 5640   5640 13.160105 -0.17659649   Male       No 1152   Public   0.41   0.256
#> 5650   5650 14.273533  0.02244444 Female      Yes  720 Catholic   0.60  -0.070
#> 5720   5720 14.282302  0.03256604   Male       No  381 Catholic   0.65  -0.352
#> 5761   5761 11.138058 -0.32300000 Female       No  215 Catholic   0.63  -0.892
#> 5762   5762  4.324865 -1.19394595 Female      Yes 1826   Public   0.24   0.364
#> 6074   6074 13.779089 -0.27675000 Female       No 2051 Catholic   0.32  -1.018
#> 6484   6484 12.912400 -0.18514286 Female       No  726   Public   0.19   0.218
#> 6897   6897 15.097633  0.34955102 Female       No 1415   Public   0.55  -0.361
#> 7172   7172  8.066818 -0.28981818 Female      Yes  280 Catholic   0.05   1.013
#> 7232   7232 12.542635 -0.09011538 Female       No 1154   Public   0.20   0.975
#> 7342   7342 11.166414 -0.44782759   Male       No 1220 Catholic   0.46   0.380
#> 7345   7345 11.338554  0.03325000 Female       No  978   Public   0.64   0.336
#> 7688   7688 18.422315  0.18588889   Male       No 1410 Catholic   0.65  -0.575
#> 7697   7697 15.721781  0.25825000   Male       No 1734   Public   0.20   0.279
#> 7890   7890  8.341098 -0.52270588   Male       No  311   Public   0.21   0.845
#> 7919   7919 14.849973  0.45767568   Male       No 1451   Public   0.50  -0.402
#> 8531   8531 13.528683  0.40809756 Female       No 2190   Public   0.58   0.132
#> 8627   8627 10.883717  0.10483019   Male       No 2452   Public   0.25   0.742
#> 8707   8707 12.883938  0.15512500 Female       No 1133   Public   0.48   1.542
#> 8854   8854  4.239781 -0.75675000 Female       No  745   Public   0.18  -0.228
#> 8874   8874 12.055028 -0.34077778 Female       No 2650   Public   0.20   1.742
#> 9550   9550 11.089138  0.05303448 Female       No 1532   Public   0.45   0.791
#> 9586   9586 14.863695  0.62115254 Female       No  262 Catholic   1.00  -2.416
#>      SexFemale   SexMale Freq
#> 1317 1.0000000 0.0000000   48
#> 1906 0.5094340 0.4905660   53
#> 2208 0.5833333 0.4166667   60
#> 2458 1.0000000 0.0000000   57
#> 2626 0.4736842 0.5263158   38
#> 2629 0.0000000 1.0000000   57
#> 2639 0.5714286 0.4285714   42
#> 2658 0.6000000 0.4000000   45
#> 2771 0.5090909 0.4909091   55
#> 3013 0.3584906 0.6415094   53
#> 3610 0.4531250 0.5468750   64
#> 3992 0.3962264 0.6037736   53
#> 4292 0.0000000 1.0000000   65
#> 4511 1.0000000 0.0000000   58
#> 4530 1.0000000 0.0000000   63
#> 4868 0.3235294 0.6764706   34
#> 5619 0.4545455 0.5454545   66
#> 5640 0.4210526 0.5789474   57
#> 5650 0.7111111 0.2888889   45
#> 5720 0.4528302 0.5471698   53
#> 5761 1.0000000 0.0000000   52
#> 5762 0.5675676 0.4324324   37
#> 6074 1.0000000 0.0000000   56
#> 6484 0.5714286 0.4285714   35
#> 6897 0.5918367 0.4081633   49
#> 7172 0.5000000 0.5000000   44
#> 7232 0.5769231 0.4230769   52
#> 7342 0.0000000 1.0000000   58
#> 7345 0.5178571 0.4821429   56
#> 7688 0.0000000 1.0000000   54
#> 7697 0.3437500 0.6562500   32
#> 7890 0.4705882 0.5294118   51
#> 7919 0.4324324 0.5675676   37
#> 8531 0.5609756 0.4390244   41
#> 8627 0.4528302 0.5471698   53
#> 8707 0.5416667 0.4583333   48
#> 8854 0.5312500 0.4687500   32
#> 8874 0.5833333 0.4166667   36
#> 9550 0.6551724 0.3448276   29
#> 9586 1.0000000 0.0000000   59

    # Similar using 'agg'
    
    up(hs, ~school, agg = ~ Sex)
#>      school Size   Sector PRACAD DISCLIM Freq Sex_Female  Sex_Male
#> 1317   1317  455 Catholic   0.95  -1.694   48  1.0000000 0.0000000
#> 1906   1906  400 Catholic   0.87  -0.939   53  0.5094340 0.4905660
#> 2208   2208 1061 Catholic   0.68  -0.864   60  0.5833333 0.4166667
#> 2458   2458  545 Catholic   0.89  -1.484   57  1.0000000 0.0000000
#> 2626   2626 2142   Public   0.40   0.142   38  0.4736842 0.5263158
#> 2629   2629 1314 Catholic   0.81  -0.613   57  0.0000000 1.0000000
#> 2639   2639 2713   Public   0.14  -0.282   42  0.5714286 0.4285714
#> 2658   2658  780 Catholic   0.79  -0.961   45  0.6000000 0.4000000
#> 2771   2771  415   Public   0.24   1.048   55  0.5090909 0.4909091
#> 3013   3013  760   Public   0.56  -0.213   53  0.3584906 0.6415094
#> 3610   3610 1431 Catholic   0.80  -0.621   64  0.4531250 0.5468750
#> 3992   3992 1114 Catholic   0.73  -1.534   53  0.3962264 0.6037736
#> 4292   4292 1328 Catholic   0.76  -0.674   65  0.0000000 1.0000000
#> 4511   4511 1068 Catholic   0.52  -1.872   58  1.0000000 0.0000000
#> 4530   4530  435 Catholic   0.60  -0.245   63  1.0000000 0.0000000
#> 4868   4868  657 Catholic   1.00  -0.219   34  0.3235294 0.6764706
#> 5619   5619 1118 Catholic   0.77  -1.286   66  0.4545455 0.5454545
#> 5640   5640 1152   Public   0.41   0.256   57  0.4210526 0.5789474
#> 5650   5650  720 Catholic   0.60  -0.070   45  0.7111111 0.2888889
#> 5720   5720  381 Catholic   0.65  -0.352   53  0.4528302 0.5471698
#> 5761   5761  215 Catholic   0.63  -0.892   52  1.0000000 0.0000000
#> 5762   5762 1826   Public   0.24   0.364   37  0.5675676 0.4324324
#> 6074   6074 2051 Catholic   0.32  -1.018   56  1.0000000 0.0000000
#> 6484   6484  726   Public   0.19   0.218   35  0.5714286 0.4285714
#> 6897   6897 1415   Public   0.55  -0.361   49  0.5918367 0.4081633
#> 7172   7172  280 Catholic   0.05   1.013   44  0.5000000 0.5000000
#> 7232   7232 1154   Public   0.20   0.975   52  0.5769231 0.4230769
#> 7342   7342 1220 Catholic   0.46   0.380   58  0.0000000 1.0000000
#> 7345   7345  978   Public   0.64   0.336   56  0.5178571 0.4821429
#> 7688   7688 1410 Catholic   0.65  -0.575   54  0.0000000 1.0000000
#> 7697   7697 1734   Public   0.20   0.279   32  0.3437500 0.6562500
#> 7890   7890  311   Public   0.21   0.845   51  0.4705882 0.5294118
#> 7919   7919 1451   Public   0.50  -0.402   37  0.4324324 0.5675676
#> 8531   8531 2190   Public   0.58   0.132   41  0.5609756 0.4390244
#> 8627   8627 2452   Public   0.25   0.742   53  0.4528302 0.5471698
#> 8707   8707 1133   Public   0.48   1.542   48  0.5416667 0.4583333
#> 8854   8854  745   Public   0.18  -0.228   32  0.5312500 0.4687500
#> 8874   8874 2650   Public   0.20   1.742   36  0.5833333 0.4166667
#> 9550   9550 1532   Public   0.45   0.791   29  0.6551724 0.3448276
#> 9586   9586  262 Catholic   1.00  -2.416   59  1.0000000 0.0000000
    
    ## To plot a summary between-cluster panel along with within-cluster panels:

    hsu <- up( hs, ~ school, all = TRUE)
    hsu$school <- ' between'  # space to make it come lexicographically before cluster names

    require( lattice )
    xyplot( mathach ~ ses | school, rbind(hs,hsu),
        panel = function( x, y, ...) {
            panel.xyplot( x, y, ...)
            panel.lmline( x, y, ...)
        } )
#> Error in rbind(deparse.level, ...): numbers of columns of arguments do not match
        
    ## To create a data frame grouped by predictors with frequency variables for each
    ## level of a response variable for analysis with a binomial glm with goodness of fit
    ## based on the deviance
    
    hsa <- up( hs, ~school, freq = ~ Sex)
#> Error in FUN(X[[i]], ...): invalid 'type' (closure) of argument
    head(hsa)
#>      school  mathach         ses    Sex Minority Size   Sector PRACAD DISCLIM
#> 1317   1317 13.17769  0.34533333 Female      Yes  455 Catholic   0.95  -1.694
#> 1906   1906 15.98317  0.51162264 Female       No  400 Catholic   0.87  -0.939
#> 2208   2208 15.40467  0.42316667 Female       No 1061 Catholic   0.68  -0.864
#> 2458   2458 13.98568  0.22778947 Female      Yes  545 Catholic   0.89  -1.484
#> 2626   2626 13.39661 -0.06484211   Male       No 2142   Public   0.40   0.142
#> 2629   2629 14.90777 -0.13764912   Male       No 1314 Catholic   0.81  -0.613
#>      Freq
#> 1317   48
#> 1906   53
#> 2208   60
#> 2458   57
#> 2626   38
#> 2629   57
    fit <- glm(cbind(Sex_Female, Sex_Male) ~ Sector, hsa, family = binomial)
#> Error in eval(predvars, data, env): object 'Sex_Female' not found
    summary(fit) # the residual deviance provides a goodness of fit test
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'object' in selecting a method for function 'summary': object 'fit' not found