Apply a function to clusters of rows in a data frame

Apply a function to clusters of rows in a data frame and return the result so it is conformable with the data frame created by 'up' applied to the same data frame and same clustering formula

Usage

up_apply(object, form, FUN, ..., sep = "/")

Arguments

object

a data frame as source for an aggregated result

form

a one-sided formula identifying the variable(s) in object that identifies clusters. e.g. ~ school/Sex to get a summary within each Sex of each school

FUN

a function to be applied to each data frame consisting of a cluster of rows of 'object'. The most common choice is with so that '...' can be an expression using variable names in 'object'.

...

other arguments to FUN, frequently when FUN is 'with', an expression using variable names in 'object'

Examples

zd <- data.frame(a=c('a','a','b','b','c','c','c'),
      b = c("B","B","A","B","C","D","D"), x = 1:7, y = 11:17)
zd$n <- capply(zd$x, zd[c('a','b')], length)
zdu <- up(zd, ~a, agg = ~b)
zdu
#>   a Freq b_A b_B       b_C       b_D
#> a a    2 0.0 1.0 0.0000000 0.0000000
#> b b    2 0.5 0.5 0.0000000 0.0000000
#> c c    3 0.0 0.0 0.3333333 0.6666667
zdu$p <- up_apply(zd, ~ a, with, sum(x)/sum(y))
zdu
#>   a Freq b_A b_B       b_C       b_D         p
#> a a    2 0.0 1.0 0.0000000 0.0000000 0.1304348
#> b b    2 0.5 0.5 0.0000000 0.0000000 0.2592593
#> c c    3 0.0 0.0 0.3333333 0.6666667 0.3750000